% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\name{adjust_policy}
\alias{adjust_policy}
\alias{adjust_policy<-}
\title{Getter/Setter for Adjust Policy}
\usage{
adjust_policy(object)

adjust_policy(object) <- value

\S4method{adjust_policy}{sim}(object)

\S4method{adjust_policy}{sim,character}(object) <- value
}
\arguments{
\item{object}{An S4 sim object.}

\item{value}{A character value that can either be \code{"back_off"} for backing off strategy or \code{"none"}.}
}
\description{
Adjust policy specifies the adjustment scheme for sequential scheduling/prediction failures. \code{"back_off"} enforeces backing off strategy that stops scheduling/predicting after a scheduling/prediction has failed, and resume scheduling/predicting after a scheuling/prediction has survived.
}
