% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_hoc_analysis.R
\name{calc_removal_to_reach_target}
\alias{calc_removal_to_reach_target}
\title{Compute the Traces to Remove in order for Score1 to Reach Target}
\usage{
calc_removal_to_reach_target(
  predict_info_quantiles,
  cut_off_prob,
  target = 1 - cut_off_prob,
  adjustment = FALSE
)
}
\arguments{
\item{predict_info_quantiles}{A dataframe storing the prediction info.}

\item{cut_off_prob}{A numeric value representing the cut off probability to considered.}

\item{target}{A numeric value for the target of score1.}

\item{adjustment}{A logical value or \code{NA} representing whether to consider adjustment. \code{NA} represents only consider adjustment when score1 is below target.}
}
\value{
A list containing adjusted \code{predict_info_quantiles} and a vector of trace_name removed.
}
\description{
Compute the traces to remove in order for score1 to reach target.
}
