% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_actual}
\alias{check_actual}
\title{Check Actual Information of Scheduling Job.}
\usage{
check_actual(actual_obs, cpu_required, granularity)
}
\arguments{
\item{actual_obs}{The actual observation of the time series.}

\item{cpu_required}{The cpu required by the job that needs to be scheduled.}

\item{granularity}{The granularity of 100 percent of total cpu.}
}
\value{
\code{0} if the job scheduled actually survives, otherwise the location the job fails.
}
\description{
Check the actual information if the job is scheduled.
}
\keyword{internal}
