% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_decision}
\alias{check_decision}
\title{Check Decision of Scheduling Job At Next Window.}
\usage{
check_decision(prob, cut_off_prob)
}
\arguments{
\item{prob}{The probability of next scheduling will fail.}

\item{cut_off_prob}{The maximum probability allowed to have next scheduling failing.}
}
\value{
\code{1} if the scheduling decision is Yes, \code{0} otherwise.
}
\description{
Make decision based on the probability and cut off probability.
}
\keyword{internal}
