% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_decision_trace_after_adjustment}
\alias{check_decision_trace_after_adjustment}
\title{Check Adjustment Status After a Specific Setting of Adjustment Policy}
\usage{
check_decision_trace_after_adjustment(score1, adjustment_policy)
}
\arguments{
\item{score1}{A numeric vector representing the vanilla score1.}

\item{adjustment_policy}{A numeric length \code{2} vector representing the adjustment policy.}
}
\value{
A numeric vector of \code{0}s and \code{1}s representing whether adjustment policy is activated.
}
\description{
Find adjustment policy after applying adjustment policy.
}
\keyword{internal}
