% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_residual}
\alias{check_residual}
\title{Check Residual Of A Single Prediction.}
\usage{
check_residual(expected, actual_obs, window_size, response)
}
\arguments{
\item{expected}{The expected value of next observation.}

\item{actual_obs}{The actual observation corresponding to the predictions.}

\item{window_size}{The length of predictions.}

\item{response}{If \code{"max"} is provided, then take max for each \code{new_freq} observations, if \code{"avg"} is provided, take avg for each \code{new_freq} observations.}
}
\value{
If expected is NA, then NA_real_ is returned, otherwise return aggregated actual subtract by expected value.
}
\description{
Check the forecast residual.
}
\keyword{internal}
