% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_score1_with_failed_pos}
\alias{check_score1_with_failed_pos}
\title{Check Score 1 Based on Prediction}
\usage{
check_score1_with_failed_pos(pi_up, actual_obs, granularity)
}
\arguments{
\item{pi_up}{A numeric value representing the prediction upper bound.}

\item{actual_obs}{A numeric value representing the actual observation corresponding to the predictions.}

\item{granularity}{A numeric value represneting granularity of response.}
}
\value{
A numeric integer or a length \code{0} integer representing the position of failure.
}
\description{
Check Score 1 Based on Prediction
}
\keyword{internal}
