% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_score2}
\alias{check_score2}
\title{Check Score 2 Based on Prediction}
\usage{
check_score2(pi_up, actual_obs, granularity, how = "max_size")
}
\arguments{
\item{pi_up}{A numeric value representing the prediction upper bound.}

\item{actual_obs}{A numeric value representing the actual observation corresponding to the predictions.}

\item{granularity}{A numeric value represneting granularity of response.}

\item{how}{A character representing how the score is calculated. \code{"max_size"} by default schedules one job that takes maximum place. \code{"\\d_jobs"} schedules \code{"\\d"} jobs with equal sizes. \code{"\\d_cores"} schedules jobs each with fixed \code{\\d} number of cores, only works for \code{granularity} not being \code{0}.}
}
\value{
A numeric value between \code{0} and \code{1} or \code{NA}.
}
\description{
Check score 2 of the prediction based on actual observation and the granularity of simulation.
}
\keyword{internal}
