% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_score_param}
\alias{check_score_param}
\title{Check Score Of An Entire Parameter Setting.}
\usage{
check_score_param(cut_off_prob, predict_info)
}
\arguments{
\item{cut_off_prob}{A numeric vector of cut off probability.}

\item{predict_info}{A dataframe representing prediction information of different traces.}
}
\value{
A sim_result object.
}
\description{
Check the score information of an entire parameter based on actual observations and predictions.
}
\keyword{internal}
