% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_score_param_simplified}
\alias{check_score_param_simplified}
\title{Check Score Of An Entire Parameter Setting Simplified.}
\usage{
check_score_param_simplified(trace_predict_info)
}
\arguments{
\item{trace_predict_info}{A dataframe representing prediction information of different traces.}
}
\value{
A sim_result object.
}
\description{
Check the score information of an entire parameter based on actual observations and predictions.
}
\keyword{internal}
