% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_score_pred}
\alias{check_score_pred}
\title{Check Scores Of A Single Prediction.}
\usage{
check_score_pred(object, predict_info, actual_obs, adjust_switch)
}
\arguments{
\item{object}{An S4 sim object.}

\item{predict_info}{The dataframe storing the prediction info}

\item{actual_obs}{The actual observation corresponding to the predictions.}

\item{adjust_switch}{A logical vector representing the status of the adjust_switch.}
}
\value{
The updated prediction information dataframe with last row modified.
}
\description{
Check the score information of a prediction based on actual observations and predictions
}
\keyword{internal}
