% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_score_pred_simplified}
\alias{check_score_pred_simplified}
\title{Check Scores Of A Single Prediction Simplified.}
\usage{
check_score_pred_simplified(
  train_iter,
  test_iter,
  predict_iter,
  predict_info,
  actual_obs
)
}
\arguments{
\item{train_iter}{A numeric number representing the number of iteration of training step, used as identifier when evaluating training performance.}

\item{test_iter}{A numeric number representing the number of iteration on testing step of the current training step.}

\item{predict_iter}{A numeric number representing the number of iteration on prediction step of the current testing step.}

\item{predict_info}{The dataframe storing the prediction info}

\item{actual_obs}{The actual observation corresponding to the predictions.}
}
\value{
The updated prediction information dataframe with last row modified.
}
\description{
Check the score information of a prediction based on actual observations and predictions
}
\keyword{internal}
