% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_score_test}
\alias{check_score_test}
\title{Check Score Of A Testing Batch.}
\usage{
check_score_test(test_predict_info, predict_info)
}
\arguments{
\item{test_predict_info}{A dataframe representing current prediction information.}

\item{predict_info}{A dataframe representing past prediction information.}
}
\value{
A list containing updated predict_info if applicable and sim_result object.
}
\description{
Check the score information of a testing batch based on actual observations and predictions.
}
\keyword{internal}
