% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_summary_statistics_trace}
\alias{check_summary_statistics_trace}
\title{Check Summary Statistics of Prediction Result}
\usage{
check_summary_statistics_trace(predict_info, granularity)
}
\arguments{
\item{predict_info}{A dataframe representing past prediction information.}

\item{granularity}{A numeric value represneting granularity of response.}
}
\value{
A dataframe of 1 row containing summary statistics.
}
\description{
Check the statistics of training set and residuals.
}
\keyword{internal}
