% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_survival}
\alias{check_survival}
\title{Check Survial Of A Single Prediction.}
\usage{
check_survival(pi_up, actual_obs, granularity)
}
\arguments{
\item{pi_up}{The prediction upper bound of next observation.}

\item{actual_obs}{The actual observation corresponding to the predictions.}

\item{granularity}{The granularity of 100 percent of total cpu.}
}
\value{
If both predicted available and actual available is smaller than \code{granularity}, \code{NA} is returned, if predicted available is smaller than or equal to actual, \code{0} is returned, otherwise, the first position that predicted available is greater than actual is returned.
}
\description{
Check the survival information of a prediction based on actual observations.
}
