% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_utilization}
\alias{check_utilization}
\title{Check Utilization Of Next Prediction.}
\usage{
check_utilization(pi_up, survival, granularity)
}
\arguments{
\item{pi_up}{The prediction upper bound of next observations.}

\item{survival}{The survival information of the prediction.}

\item{granularity}{The granularity of 100 percent of total cpu.}
}
\value{
A vector same size as the input vector.
}
\description{
Check the utilization of next predictions.
}
