% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovatree_pred.R
\name{check_valid_anovatree_pred}
\alias{check_valid_anovatree_pred}
\title{Validity Checker for anovatree_pred Object}
\usage{
check_valid_anovatree_pred(object)
}
\arguments{
\item{object}{A anovatree_pred object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for anovatree_pred Object
}
\keyword{internal}
