% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari11_sim.R
\name{check_valid_ari11_sim}
\alias{check_valid_ari11_sim}
\title{Validity Checker for ari11_sim Object}
\usage{
check_valid_ari11_sim(object)
}
\arguments{
\item{object}{A ar1_sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for ari11_sim Object
}
\keyword{internal}
