% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_arima_sim.R
\name{check_valid_auto_arima_sim}
\alias{check_valid_auto_arima_sim}
\title{Validity Checker for auto_arima_sim Object}
\usage{
check_valid_auto_arima_sim(object)
}
\arguments{
\item{object}{A auto_arima_sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for auto_arima_sim Object
}
\keyword{internal}
