% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopilot_sim.R
\name{check_valid_autopilot_sim}
\alias{check_valid_autopilot_sim}
\title{Validity Checker for autopilot_sim Object}
\usage{
check_valid_autopilot_sim(object)
}
\arguments{
\item{object}{A autopilot_sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for autopilot_sim Object
}
\keyword{internal}
