% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ets_sim.R
\name{check_valid_ets_sim}
\alias{check_valid_ets_sim}
\title{Validity Checker for ets_sim Object}
\usage{
check_valid_ets_sim(object)
}
\arguments{
\item{object}{A ets_sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for ets_sim Object
}
\keyword{internal}
