% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_pred.R
\name{check_valid_gmm_pred}
\alias{check_valid_gmm_pred}
\title{Validity Checker for gmm_pred Object}
\usage{
check_valid_gmm_pred(object)
}
\arguments{
\item{object}{A gmm_pred object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for gmm_pred Object
}
\keyword{internal}
