% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_sim.R
\name{check_valid_lm_sim}
\alias{check_valid_lm_sim}
\title{Validity Checker for lm_sim Object}
\usage{
check_valid_lm_sim(object)
}
\arguments{
\item{object}{A lm_sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for lm_sim Object
}
\keyword{internal}
