% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_sim.R
\name{check_valid_markov_sim}
\alias{check_valid_markov_sim}
\title{Validity Checker for markov_sim Object}
\usage{
check_valid_markov_sim(object)
}
\arguments{
\item{object}{A markov_sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for markov_sim Object
}
\keyword{internal}
