% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_sim.R
\name{check_valid_multinom_sim}
\alias{check_valid_multinom_sim}
\title{Validity Checker for multinom_sim Object}
\usage{
check_valid_multinom_sim(object)
}
\arguments{
\item{object}{A multinom_sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for multinom_sim Object
}
\keyword{internal}
