% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_sim.R
\name{check_valid_nn_sim}
\alias{check_valid_nn_sim}
\title{Validity Checker for nn_sim Object}
\usage{
check_valid_nn_sim(object)
}
\arguments{
\item{object}{A nn_sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for nn_sim Object
}
\keyword{internal}
