% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_class.R
\name{check_valid_pred}
\alias{check_valid_pred}
\title{Validity Checker for pred Object}
\usage{
check_valid_pred(object)
}
\arguments{
\item{object}{A pred object}
}
\value{
\code{TRUE} if the input pred object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for pred Object
}
\keyword{internal}
