% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_class.R
\name{check_valid_sim}
\alias{check_valid_sim}
\title{Validity Checker for sim Object}
\usage{
check_valid_sim(object)
}
\arguments{
\item{object}{A sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for sim Object
}
\keyword{internal}
