% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlm_sim.R
\name{check_valid_stlm_sim}
\alias{check_valid_stlm_sim}
\title{Validity Checker for stlm_sim Object}
\usage{
check_valid_stlm_sim(object)
}
\arguments{
\item{object}{A stlm_sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for stlm_sim Object
}
\keyword{internal}
