% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structal_sim.R
\name{check_valid_structual_sim}
\alias{check_valid_structual_sim}
\title{Validity Checker for structual_sim Object}
\usage{
check_valid_structual_sim(object)
}
\arguments{
\item{object}{A structual_sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for structual_sim Object
}
\keyword{internal}
