% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surtree_pred.R
\name{check_valid_surtree_pred}
\alias{check_valid_surtree_pred}
\title{Validity Checker for surtree_pred Object}
\usage{
check_valid_surtree_pred(object)
}
\arguments{
\item{object}{A surtree_pred object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for surtree_pred Object
}
\keyword{internal}
