% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var1_sim.R
\name{check_valid_var1_sim}
\alias{check_valid_var1_sim}
\title{Validity Checker for var1_sim Object}
\usage{
check_valid_var1_sim(object)
}
\arguments{
\item{object}{A var1_sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for var1_sim Object
}
\keyword{internal}
