% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_sim.R
\name{check_valid_var_sim}
\alias{check_valid_var_sim}
\title{Validity Checker for var_sim Object}
\usage{
check_valid_var_sim(object)
}
\arguments{
\item{object}{A var_sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for var_sim Object
}
\keyword{internal}
