% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vari11_sim.R
\name{check_valid_vari11_sim}
\alias{check_valid_vari11_sim}
\title{Validity Checker for vari11_sim Object}
\usage{
check_valid_vari11_sim(object)
}
\arguments{
\item{object}{A vari11_sim object}
}
\value{
\code{TRUE} if the input sim object is valid, vector of error messages otherwise.
}
\description{
Validity Checker for vari11_sim Object
}
\keyword{internal}
