% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_arguments.R
\name{check_values}
\alias{check_values}
\title{Check Input Parameters.}
\usage{
check_values(param, value)
}
\arguments{
\item{param}{A String that is either \code{window_size}, \code{cut_off_prob}, \code{granularity}, \code{train_size} or \code{update_freq}.}

\item{value}{A value that the parameter is about to be assigned to.}
}
\description{
Check whether the value satisfies the precondition of the parameter.
}
