% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{combine_result}
\alias{combine_result}
\title{Combine Two Simulation Result.}
\usage{
combine_result(object1, object2)
}
\arguments{
\item{object1}{A sim_result object to be combined with another.}

\item{object2}{A sim_result object to be combined with another.}
}
\value{
A sim_result object with type being the higher level of the two objects being combined.
}
\description{
Combine two sim_result object and return the combined object.
}
\keyword{internal}
