% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{combined_sim_job_centered}
\alias{combined_sim_job_centered}
\title{Combinations of Predictions of Background Jobs and Foreground Jobs}
\usage{
combined_sim_job_centered(
  load_foreground_result = NULL,
  param_setting_sim = NULL,
  additional_param_sim = list(),
  foreground_x = NULL,
  foreground_xreg = NULL,
  sim_length = 200,
  load_background_result = NULL,
  param_setting_pred = NULL,
  additional_param_pred = list(),
  background_x = NULL,
  background_xreg = NULL,
  bins = c(0, 1, 2, 6, 10, 14, 18, 22, 26, 30, 50, 80, 205),
  cores = parallel::detectCores(),
  write_type = "none",
  result_loc = getwd()
)
}
\arguments{
\item{load_foreground_result}{A character value representing the location of the previously run foregound simulation results. Default value is \code{NULL}.}

\item{param_setting_sim}{A dataframe representing a specific parameter setting for sim object.}

\item{additional_param_sim}{A list containing additional arguments to be passed into param_sim.}

\item{foreground_x}{A matrix of size n by m representing the target dataset for scheduling and evaluations.}

\item{foreground_xreg}{A matrix of size n by m representing the target dataset for scheduling and evaluations, or \code{NULL}.}

\item{sim_length}{A numeric integer representing the length of time for simulation, training size excluded.}

\item{load_background_result}{A character value representing the location of the previously run foregound simulation results. Default value is \code{NULL}.}

\item{param_setting_pred}{A dataframe representing a specific parameter setting for pred object.}

\item{additional_param_pred}{A list containing additional arguments to be passed into param_pred.}

\item{background_x}{A matrix of size n by m representing the target dataset for scheduling and evaluations.}

\item{background_xreg}{A matrix of size n by m representing the dataset that target dataset depends on for predicting.}

\item{bins}{A numeric vector representing the discretization will be used on background job length, the first value representing the lower bound such as \code{0}, last value representing the upper bound such as \code{205}.}

\item{cores}{A numeric value representing the number of threads for parallel programming for multiple traces, not supported for windows users.}

\item{write_type}{A character that represents how to write the result of simulation, can be "detailed", "summary" or "none". Default value is \code{"none"}.}

\item{result_loc}{A character that specify the path to which the result of simulations will be saved to. Default is your work directory.}

\item{constraint_prob}{A numeric value representing the cut off for survival probability of a job to be scheduled.}

\item{heartbeats_percent}{A numeric value representing the percentage of sampled machines will periodically send "heat beats", which is the availability information to the scheduler. Default value is \code{1}.}

\item{machines_full_indicator}{An numeric value that is used to identify whether all machines are full. The scheduler believes the machines are full after \code{machines_full_indicator} successive failures in scheduling.}
}
\value{
A dataframe containing the decisions made by scheduler.
}
\description{
Sequentially training and testing by predicting the availability of CPU resource at next windows.
}
