% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{compute_performance}
\alias{compute_performance}
\title{Compute Overall Performance of Predictions.}
\usage{
compute_performance(predictions, actuals)
}
\arguments{
\item{predictions}{The predictions made, a vector of \code{0}, \code{1} and \code{NA}.}

\item{actuals}{The actual survival information, a vector of \code{0} and \code{1}.}
}
\value{
A list consists scheduled num, unscheduled num, correctly scheduled num, correctly unscheduled num.
}
\description{
Compute overall correctly scheduled ratio and correctlt unscheduled ratio.
}
\keyword{internal}
