% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{compute_pi_up}
\alias{compute_pi_up}
\title{Compute Prediction Interval Upper Bound}
\usage{
compute_pi_up(object)
}
\arguments{
\item{object}{An S4 sim process object with prediction information.}
}
\value{
A numeric value representing the maximum of prediction interval upper bound in the forecasting steps.
}
\description{
This is a generic function that computes the upper bound of prediction interval according to the type of input object.
}
