% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{compute_pi_up_markov}
\alias{compute_pi_up_markov}
\title{Computation of PI's upper bound for Markov Model}
\usage{
compute_pi_up_markov(to_states, prob_cut_off, granularity)
}
\arguments{
\item{to_states}{The vector contains the probability of going to each state}

\item{prob_cut_off}{The probability cut off point for PI}

\item{granularity}{The granularity of 100 percent of total cpu.}
}
\value{
The prediction upper bound.
}
\description{
Compute the PI's upper bound based on probability cut off
}
\keyword{internal}
