% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{compute_summary_performance}
\alias{compute_summary_performance}
\title{Compute Performance of Scheduling of Jobs in Summary}
\usage{
compute_summary_performance(
  predict_info,
  past_failures,
  machine_available_resources,
  sim_end_time,
  window_multiplier
)
}
\arguments{
\item{predict_info}{A dataframe containing all the jobs scheduling information.}

\item{past_failures}{A dataframe containing all the jobs killed by more than once.}

\item{machine_available_resources}{A numeric number representing all the available resources across all machines of all time.}

\item{sim_end_time}{A numeric number representing the finish time of simulation.}

\item{window_multiplier}{A numeric number representing the multiplier for delayed time and scheduled time.}
}
\value{
A list containing utilization for finished jobs and utilization for total jobs.
}
\description{
Compute Performance of Scheduling of Jobs in Summary
}
