% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{compute_survival}
\alias{compute_survival}
\title{Compute The Overall Survial.}
\usage{
compute_survival(survival)
}
\arguments{
\item{survival}{A vector of survial information.}
}
\value{
A list containing the information of overall survival information.
}
\description{
Compute the overall survival rate given survival informations.
}
\keyword{internal}
