% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{compute_utilization}
\alias{compute_utilization}
\title{Compute The Overall Utilization.}
\usage{
compute_utilization(utilization, actual_obs, window_size, granularity)
}
\arguments{
\item{utilization}{A vector of survial information.}

\item{actual_obs}{A vector of actual observations.}

\item{window_size}{The length of predictions.}

\item{granularity}{The granularity of 100 percent of total cpu.}
}
\value{
A list containing the information of overall survival information.
}
\description{
Compute the overall utilization rate given utilization informations.
}
