% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{convert_frequency_dataset_overlapping}
\alias{convert_frequency_dataset_overlapping}
\title{Convert Dataset Frequency Using Overlapping Windows.}
\usage{
convert_frequency_dataset_overlapping(dataset, new_freq, mode)
}
\arguments{
\item{dataset}{A vector of numeric value.}

\item{new_freq}{An integer value.}

\item{mode}{If \code{"max"} is provided, then take max for each \code{new_freq} observations, if \code{"avg"} is provided, take avg for each \code{new_freq} observations.}
}
\value{
The vector of same size of input vector.
}
\description{
Convert \code{dataset} to bigger frequency provided by \code{new_freq} using max/avg operators, the rest will be truncated off.
}
\keyword{internal}
