% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\name{cut_off_prob}
\alias{cut_off_prob}
\alias{cut_off_prob<-}
\title{Getter/Setter for Cut Off Probability}
\usage{
cut_off_prob(object)

cut_off_prob(object) <- value

\S4method{cut_off_prob}{sim}(object)

\S4method{cut_off_prob}{sim,numeric}(object) <- value
}
\arguments{
\item{object}{An S4 sim object}

\item{value}{A numeric value between 0 and 1, exclusive.}
}
\description{
A cut off probability is the minimum probability allowed to have next observation below a input threshold, or equals to \eqn{\frac{1}{2}(1 - Prob)} as \eqn{Prob} stands for Probability in Prediction Interval.
}
