% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{discretization}
\alias{discretization}
\title{Discretize A Vector of Duration Time.}
\usage{
discretization(breakpoints, vec)
}
\arguments{
\item{breakpoints}{A numeric vector specifies how do these times discretized into bins.}

\item{vec}{A numeric vector of time to be discretized.}
}
\description{
Discretize a vector of duration time according to specified bins.
}
\keyword{internal}
