% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ar1_markov_sim.R, R/ar1_sim.R,
%   R/ari11_sim.R, R/markov_sim.R, R/var1_sim.R, R/vari11_sim.R
\name{do_prediction}
\alias{do_prediction}
\alias{do_prediction,ar1_markov_sim,list,numeric,numeric,numeric-method}
\alias{do_prediction,ar1_sim,list,numeric,numeric,numeric-method}
\alias{do_prediction,ari11_sim,list,numeric,numeric,numeric-method}
\alias{do_prediction,markov_sim,list,numeric,numeric,numeric-method}
\alias{do_prediction,var1_sim,list,numeric,numeric,numeric-method}
\alias{do_prediction,vari11_sim,list,numeric,numeric,numeric-method}
\title{Do Prediction}
\usage{
do_prediction(object, trained_result, last_obs_max, last_obs_avg, level)

\S4method{do_prediction}{ar1_markov_sim,list,numeric,numeric,numeric}(object, trained_result, last_obs_max, last_obs_avg, level)

\S4method{do_prediction}{ar1_sim,list,numeric,numeric,numeric}(object, trained_result, last_obs_max, last_obs_avg, level)

\S4method{do_prediction}{ari11_sim,list,numeric,numeric,numeric}(object, trained_result, last_obs_max, last_obs_avg, level)

\S4method{do_prediction}{markov_sim,list,numeric,numeric,numeric}(object, trained_result, last_obs_max, last_obs_avg, level)

\S4method{do_prediction}{var1_sim,list,numeric,numeric,numeric}(object, trained_result, last_obs_max, last_obs_avg, level)

\S4method{do_prediction}{vari11_sim,list,numeric,numeric,numeric}(object, trained_result, last_obs_max, last_obs_avg, level)
}
\arguments{
\item{object}{An S4 sim process object}

\item{trained_result}{A list containing trained model information.}

\item{last_obs_max}{A numeric value representing the last observation of maximum.}

\item{last_obs_avg}{A numeric value representing the last observation of average.}

\item{level}{The level in \eqn{Pr(Y_{t+1}|Y_{t}) \leq level} to be computed.}
}
\value{
Prediction information.
}
\description{
This is a generic function that do prediction according to the input object type.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = ar1_markov_sim,trained_result = list,last_obs_max = numeric,last_obs_avg = numeric,level = numeric}: Do prediction based on trained AR1-Markov Model.

\item \code{object = ar1_sim,trained_result = list,last_obs_max = numeric,last_obs_avg = numeric,level = numeric}: Do prediction based on trained AR1 Model.

\item \code{object = ari11_sim,trained_result = list,last_obs_max = numeric,last_obs_avg = numeric,level = numeric}: Do prediction based on trained ARI Model.

\item \code{object = markov_sim,trained_result = list,last_obs_max = numeric,last_obs_avg = numeric,level = numeric}: Do prediction based on trained Markov Model.

\item \code{object = var1_sim,trained_result = list,last_obs_max = numeric,last_obs_avg = numeric,level = numeric}: Do prediction based on trained VAR Model.

\item \code{object = vari11_sim,trained_result = list,last_obs_max = numeric,last_obs_avg = numeric,level = numeric}: Do prediction based on trained VAR Model.
}}

