% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_logistic_sim.R
\name{do_prediction_ar1_logistic}
\alias{do_prediction_ar1_logistic}
\title{Predict Next Observation For AR1-Logistic Model}
\usage{
do_prediction_ar1_logistic(last_obs, trained_result, predict_size, level)
}
\arguments{
\item{last_obs}{The previous observation.}

\item{trained_result}{The trained result from function \code{train_ar1_logistic}}

\item{predict_size}{The number of steps to predict forward.}

\item{level}{The level in \eqn{Pr(next_obs \leq level)} if the probability is not needed.}
}
\value{
A list containing the calculated probability
}
\description{
Compute \eqn{Pr(next_obs \leq level)} if \code{level} is provided, otherwise, compurte \eqn{E[next_obs|prev_obs]} and \eqn{Var[next_obs|prev_obs]}.
}
\keyword{internal}
