% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_sim.R
\name{do_prediction_markov}
\alias{do_prediction_markov}
\title{Predict Next Observation For Markov Model}
\usage{
do_prediction_markov(last_obs, trained_result, predict_size, level = NULL)
}
\arguments{
\item{last_obs}{The last observation from which the prediction is carried on.}

\item{trained_result}{The transition matrix trained for the Markov Model.}

\item{predict_size}{The number of steps to predict forward.}

\item{level}{The level in \eqn{Pr(next_obs \leq level)}, or \code{NULL} if the probability is not needed.}
}
\value{
A list containing the calculated probability
}
\description{
Compute \eqn{Pr(next_obs \leq level)} if \code{level} is provided, otherwise, compurte \eqn{E[next_obs|prev_obs]} and \eqn{Var[next_obs|prev_obs]}.
}
