% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.sim.R
\name{do_prediction_var}
\alias{do_prediction_var}
\title{Predict Next Observation For VAR Model}
\usage{
do_prediction_var(
  intercept,
  ma_coef,
  ar_coef,
  p,
  q,
  var,
  current_err,
  last_obs,
  ts_model,
  predict_size = 1,
  prev,
  level = NULL
)
}
\arguments{
\item{last_obs}{The previous observation.}

\item{predict_size}{The number of steps to predict forward.}

\item{level}{The level in \eqn{Pr(next_obs \leq level)}, or \code{NULL} if the probability is not needed.}

\item{phi}{The slope in \eqn{(Y_t - \mu) = \phi (Y_{t-1} - \mu) + e_t} where \eqn{e_t} is white noise process.}

\item{mean}{The mean of the stationary process with \eqn{\mu = E[Y_t]}.}

\item{variance}{The variance of residuals, used to estimate the variance of the error term, which is a white noise process of zero mean.}
}
\value{
A list containing the calculated probability, expectation and variance.
}
\description{
Compute \eqn{Pr(next_obs \leq level)} if \code{level} is provided, otherwise, compurte \eqn{E[next_obs|prev_obs]} and \eqn{Var[next_obs|prev_obs]}.
}
