% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var1_sim.R
\name{do_prediction_var1}
\alias{do_prediction_var1}
\title{Predict Next Observation For VAR Model}
\usage{
do_prediction_var1(last_obs, ts_model, predict_size, level = NULL)
}
\arguments{
\item{last_obs}{The previous observation.}

\item{ts_model}{A list get from training var model}

\item{predict_size}{The number of steps to predict forward.}

\item{level}{The level in \eqn{Pr(next_obs \leq level)}, or \code{NULL} if the probability is not needed.}
}
\value{
A list containing the calculated probability, expectation and variance.
}
\description{
Compute \eqn{Pr(next_obs \leq level)} if \code{level} is provided, otherwise, compurte \eqn{E[next_obs|prev_obs]} and \eqn{Var[next_obs|prev_obs]}.
}
