% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_sim.R
\name{estimate_outliers}
\alias{estimate_outliers}
\title{Estimate the effect of outliers}
\usage{
estimate_outliers(object, ol_type, new_train_x, tso_model, trained_model)
}
\arguments{
\item{object}{A arima_sim object.}

\item{ol_type}{A vector of outlier types to be considered.}

\item{new_train_x}{A time series representing the train data.}

\item{tso_model}{A tso object from training data}

\item{trained_model}{A list containing previous trained model.}
}
\value{
A list containing the means and variance of outliers for each type of outliers.
}
\description{
Estimate the effect of outliers
}
\keyword{internal}
