% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_sim.R
\name{find_and_remove_outliers_testing_set}
\alias{find_and_remove_outliers_testing_set}
\title{Find and Adjust the Newly Observed Data in Tesing Set}
\usage{
find_and_remove_outliers_testing_set(object, trained_result, new_x, res)
}
\arguments{
\item{object}{A arima_sim object.}

\item{trained_result}{A tso object.}

\item{new_x}{A time series of new observations.}

\item{res}{A time series of residuals up to current time.}
}
\value{
A time series of new observations with outlier effect adjusted.
}
\description{
Find and Adjust the Newly Observed Data in Tesing Set
}
\keyword{internal}
