% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{find_best_candidate}
\alias{find_best_candidate}
\title{Find Best Performed Model In Candidate Models.}
\usage{
find_best_candidate(candidate_model, predict_histories)
}
\arguments{
\item{candidate_model}{A numeric vector representing the pool from which the best candidate model will be selected.}

\item{predict_results}{A list of sim_result objects representing the prediction histories of models.}
}
\value{
The index of best performance candidate model.
}
\description{
Find best performed model based on each model performance history.
}
\keyword{internal}
