% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{find_expectation_gaussian_mixture}
\alias{find_expectation_gaussian_mixture}
\title{Find Expected Value of Gaussian Mixture Random Variable}
\usage{
find_expectation_gaussian_mixture(mean, sd, pro)
}
\arguments{
\item{mean}{A numeric vector of means.}

\item{sd}{A numeric vector of standard deviations.}

\item{pro}{A numeric vector of probabilities.}
}
\value{
The corresponding expectation.
}
\description{
Find the expectation mixture of gaussians.
}
