% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{find_expectation_skewnorm}
\alias{find_expectation_skewnorm}
\title{Find Expected Value of Skew Normal Random Variable}
\usage{
find_expectation_skewnorm(xi, omega, alpha)
}
\arguments{
\item{xi}{A numeric value of parameter xi.}

\item{omega}{A numeric value of parameter omega.}

\item{alpha}{A numeric value of parameter alpha.}
}
\value{
The corresponding expectation.
}
\description{
Find the expectation of state based random variable.
}
