% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{find_expectation_state_based_dist}
\alias{find_expectation_state_based_dist}
\title{Find Expected Value of Discrete Random variable}
\usage{
find_expectation_state_based_dist(prob_dist, quantiles = NULL)
}
\arguments{
\item{prob_dist}{A numeric vector representing the probability distribution at result time.}

\item{quantiles}{A numeric vector representing quantile of partitioning training set.}
}
\value{
The corresponding expectation.
}
\description{
Find the expectation of state based random variable.
}
