% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{find_overall_evaluation}
\alias{find_overall_evaluation}
\title{Find Overall Evaluation}
\usage{
find_overall_evaluation(numerator1, denominator1, numerator2, denominator2)
}
\arguments{
\item{numerator1}{The numerator of score 1.}

\item{denominator1}{The denominator of score 1.}

\item{numerator2}{The numerator of score 2.}

\item{denominator2}{The denominator of score 2.}
}
\value{
A list consists of average and aggregated score 1 and score 2.
}
\description{
Find the overall evaluation after an epoche is completed.
}
\keyword{internal}
