% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_hoc_analysis.R
\name{find_score_after_buffer}
\alias{find_score_after_buffer}
\title{Find the Score Changes After Adding Buffers}
\usage{
find_score_after_buffer(
  cut_off_prob,
  window_size,
  granularity,
  max_cores,
  adjustment_policy,
  cores = 1,
  result_path
)
}
\arguments{
\item{cut_off_prob}{A numeric vector representing the cut off probabilities.}

\item{window_size}{A numeric vector representing the window sizes.}

\item{granularity}{A numeric value representing the granularity.}

\item{max_cores}{A numeric value representing the maximum numer of cores being considered.}

\item{adjustment_policy}{A list of numeric vectors of length 2.}

\item{cores}{A numeric value representing the number of cores used for multiprocessing.}

\item{result_path}{A string for the path where simulation results are stored.}
}
\description{
Find the Score1 and Score2 change after appending a specific size of buffers.
}
