% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_hoc_analysis.R
\name{find_score_change}
\alias{find_score_change}
\title{Find the Score Change Information}
\usage{
find_score_change(
  cut_off_prob,
  target,
  window_size,
  granularity,
  result_path,
  adjustment = FALSE
)
}
\arguments{
\item{cut_off_prob}{A vector of cut off probabilities as in y axis.}

\item{target}{A vector of targets as in x axis.}

\item{window_size}{A vector of window sizes in legend.}

\item{granularity}{A numeric value of granularity.}

\item{result_path}{A string for the path where simulation results are stored.}

\item{adjustment}{A logical value representing whether adjustment is accounted.}
}
\value{
A list with paste0(cut_off_prob, ",", target) as keys, and score change as values.
}
\description{
Find the score change dataframe from the stored result folder.
}
