% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{find_shifted_skewnorm}
\alias{find_shifted_skewnorm}
\title{Find Updated Parameters For Skew Normal Distribution}
\usage{
find_shifted_skewnorm(shift, xi, omega, alpha)
}
\arguments{
\item{shift}{A numeric value representing the constant mean shift.}

\item{xi}{A numeric value of parameter xi.}

\item{omega}{A numeric value of parameter omega.}

\item{alpha}{A numeric value of parameter alpha.}
}
\value{
The corresponding expectation.
}
\description{
Find the updated parameters for skew normal distribution after a mean shift.
}
