% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{find_shifted_state_based_dist}
\alias{find_shifted_state_based_dist}
\title{Find Discrete Random Variable Distribution After Mean Shift}
\usage{
find_shifted_state_based_dist(shift, prob_dist, quantiles = NULL)
}
\arguments{
\item{shift}{A numeric value representing the constant mean shift.}

\item{prob_dist}{A numeric vector representing the probability distribution at result time.}

\item{quantiles}{A numeric vector representing quantile of partitioning training set.}
}
\value{
The corresponding expectation.
}
\description{
Find resulting distribution after a mean shift for discrete random variable.
}
