% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{find_state_based_cdf}
\alias{find_state_based_cdf}
\title{Find CDF of Discrete Random Variable}
\usage{
find_state_based_cdf(q, prob_dist, quantiles = NULL, type = "fixed")
}
\arguments{
\item{q}{A numeric quantile.}

\item{prob_dist}{A numeric vector representing the probability distribution at result time.}

\item{quantiles}{A numeric vector representing quantile of partitioning training set.}

\item{type}{A character that can either be \code{"fixed"} or \code{"quantile"}.}
}
\value{
The corresponding probability.
}
\description{
Find the CDF of state based random variable.
}
