% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{find_worst_candidate}
\alias{find_worst_candidate}
\title{Find Worst Performed Model In Candidate Models.}
\usage{
find_worst_candidate(model_num, predict_histories)
}
\arguments{
\item{model_num}{A numeric number representing the total number of models to keep and switch within.}

\item{predict_histories}{A list of sim_result objects representing the prediction histories of models.}
}
\value{
The index of worst performance candidate model.
}
\description{
Find worst performed model based on each model performance history to be replaced by new model.
}
\keyword{internal}
