% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_arguments.R
\name{generate_parameters_df}
\alias{generate_parameters_df}
\title{Generate Dataframe of Parameters}
\usage{
generate_parameters_df(model_name)
}
\arguments{
\item{model_name}{A String that is either \code{window_size}, \code{cut_off_prob}, \code{granularity}, \code{train_size} or \code{update_freq}.}
}
\value{
A dataframe of the combination of the parameters.
}
\description{
Generate the dataframe combination of the parameters for simulation of a specific model.
}
\examples{
generate_parameters_df("AR1")
}
