% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generate_result}
\alias{generate_result}
\title{Generate The Result Of Simulation}
\usage{
generate_result(object, evaluation, write_result)
}
\arguments{
\item{object}{An S4 sim object.}

\item{evaluation}{The evaluation dataframe with each row representing each trace, and the columns consists of performance information.}

\item{write_result}{A logical TRUE/FALSE argument to determine whether to store the result of simulation to a file to location stored as an attribute in sim object.}
}
\value{
An S4 sim result object with evaluation information.
}
\description{
This is a generic function that generates of the output of simulation.
}
