% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{get_adjust_switch}
\alias{get_adjust_switch}
\title{Get Updated Adjustment Status.}
\usage{
get_adjust_switch(score1, react_counter, adjust_switch, react_speed)
}
\arguments{
\item{score1}{A numeric indicating the computed score1 information from previous prediction step, should be \code{1}, \code{0} or \code{NA}.}

\item{react_counter}{A numeric indicating the status of counter for changing the status of adjust_switch.}

\item{adjust_switch}{A logical indicating the status of adjustment switch, \code{TRUE} for on and \code{FALSE}.}

\item{react_speed}{A numeric vector indicating the reacting speed to change adjustment switch from \code{FALSE} to \code{TRUE} and from \code{TRUE} to \code{FALSE}.}
}
\value{
A list containing update step and switch information.
}
\description{
Return the updated adjustment status based on the \code{react_speed} and \code{react_counter}.
}
\keyword{internal}
