% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\name{get_numeric_slots}
\alias{get_numeric_slots}
\alias{get_numeric_slots,sim-method}
\title{Get The Numeric Parameters}
\usage{
get_numeric_slots(object)

\S4method{get_numeric_slots}{sim}(object)
}
\arguments{
\item{object}{An S4 sim object}
}
\value{
A list containing all numeric parameter informations.
}
\description{
Get The Numeric Parameters
}
