% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{get_predicting_step}
\alias{get_predicting_step}
\title{Get Update Step For Predicting.}
\usage{
get_predicting_step(
  survival,
  window_size,
  adjust_policy,
  adjust_switch,
  schedule_policy
)
}
\arguments{
\item{survival}{The survival information of current scheduling survives.}

\item{window_size}{The length of predictions.}

\item{adjust_policy}{The adjustment policy, \code{"back_off"} or \code{"none"}.}

\item{adjust_switch}{The previous switch, \code{TRUE} for on and \code{FALSE}.}

\item{schedule_policy}{The schedule policy, \code{"dynamic"} or \code{"disjoint"}.}
}
\value{
A list containing update step and switch information.
}
\description{
Get predicting update step given survival informations.
}
\keyword{internal}
