% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pre_methods.R, R/sim_methods.R
\name{get_representation}
\alias{get_representation}
\alias{get_representation,pred,character-method}
\alias{get_representation,sim,character-method}
\title{Get Representation}
\usage{
get_representation(object, type)

\S4method{get_representation}{pred,character}(object, type)

\S4method{get_representation}{sim,character}(object, type)
}
\arguments{
\item{object}{An S4 sim object.}

\item{type}{A character representing the different type of representation to be returned.}
}
\value{
A character representation of \code{object}.
}
\description{
This is a generic function that return a character representation according to the input object type.
}
