% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_arguments.R
\name{get_result_location}
\alias{get_result_location}
\title{Check Result Location}
\usage{
get_result_location()
}
\value{
A string representation of the location of the folder.
}
\description{
Return the location for the result of the simulations to be written to.
}
