% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{get_scheduling_step}
\alias{get_scheduling_step}
\title{Get Update Step For Schedulings.}
\usage{
get_scheduling_step(prediction, actual, window_size, adjust_policy,
  adjust_switch, schedule_policy)
}
\arguments{
\item{prediction}{The prediction of current scheduling.}

\item{actual}{The actual information of whether the current scheduling survives.}

\item{window_size}{The length of predictions.}

\item{adjust_policy}{The adjustment policy, \code{TRUE} or \code{FALSE}.}

\item{adjust_switch}{The previous switch, \code{TRUE} for on and \code{FALSE}.}

\item{schedule_policy}{The schedule policy, \code{"dynamic"} or \code{"disjoint"}.}
}
\value{
A list containing update step, (adjusted) prediction, and switch information.
}
\description{
Get update scheduling step given prediction and actual informations.
}
\keyword{internal}
