% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ar1_sim.R, R/markov_sim.R,
%   R/var1_sim.R
\docType{methods}
\name{get_sim_save}
\alias{get_sim_save}
\alias{get_sim_save,ar1_sim,data.frame,logical-method}
\alias{get_sim_save,markov_sim,data.frame,logical-method}
\alias{get_sim_save,var1_sim,data.frame,logical-method}
\title{Generate Simulation Result}
\usage{
get_sim_save(object, evaluation, write_result)

\S4method{get_sim_save}{ar1_sim,data.frame,logical}(object, evaluation,
  write_result)

\S4method{get_sim_save}{markov_sim,data.frame,logical}(object, evaluation,
  write_result)

\S4method{get_sim_save}{var1_sim,data.frame,logical}(object, evaluation,
  write_result)
}
\arguments{
\item{object}{An S4 sim_result object.}

\item{evaluation}{The evaluation dataframe with each row representing each trace, and the columns consists of performance information.}

\item{write_result}{A logical TRUE/FALSE argument to determine whether to store the result of simulation to a file to location stored as an attribute in sim object.}
}
\value{
An S4 sim_result object.
}
\description{
Generate Simulation Result
}
\section{Methods (by class)}{
\itemize{
\item \code{object = ar1_sim,evaluation = data.frame,write_result = logical}: Generate ar1_sim_result object from simulation.

\item \code{object = markov_sim,evaluation = data.frame,write_result = logical}: Generate markov_sim_result object from simulation.

\item \code{object = var1_sim,evaluation = data.frame,write_result = logical}: Generate ar1_sim_result object from simulation.
}}

