% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{get_training_step}
\alias{get_training_step}
\title{Get Training Update Step.}
\usage{
get_training_step(
  train_policy,
  tolerance,
  prev_score1,
  prev_score2,
  last_score1,
  last_score2
)
}
\arguments{
\item{train_policy}{The training policy, either \code{"once"}, \code{"fixed"} or \code{"dynamic"}.}

\item{tolerance}{The tolerance level of retrain, the quantile of previous performance.}

\item{prev_score1}{A vector of previous correctly scheduled rate or survival rate.}

\item{prev_score2}{A vector of previous correctly unscheduled rate or utiliztion rate.}

\item{last_score1}{The current correctly scheduled rate or survival rate.}

\item{last_score2}{The current correctly unscheduled rate or utilization rate.}
}
\value{
train signal whether \code{TRUE} if retraining is needed at next update, otherwise \code{FALSE}.
}
\description{
Computer training step.
}
\keyword{internal}
