% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{google_runtime_data}
\alias{google_runtime_data}
\title{11409 Samples of Jobs of Google ClusterData 2011 traces}
\format{
A dataframe with 11409 rows and 11 columns.
}
\source{
\url{https://github.com/google/cluster-data/blob/master/ClusterData2011_2.md}
}
\usage{
google_runtime_data
}
\description{
A dataframe conataining job wise features and running time for each row.
The running time of jobs is aggregated in 5 minutes.
}
\keyword{datasets}
