% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{google_runtime_data2}
\alias{google_runtime_data2}
\title{10000 Samples of Jobs of Google ClusterData 2011 traces}
\format{
A dataframe with 10000 rows and 11 columns.
}
\source{
\url{https://github.com/google/cluster-data/blob/master/ClusterData2011_2.md}
}
\usage{
google_runtime_data2
}
\description{
A dataframe conataining job wise features and running time for each row.
The running time of jobs is aggregated in 5 minutes.
The difference between \code{google_runtime_data} and \code{google_runtime_data2} is that \code{google_runtime_data2} is biased selection of jobs by removing
the jobs with task duration less or equal to 1.
}
\keyword{datasets}
