% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\docType{methods}
\name{granularity}
\alias{granularity}
\alias{granularity<-}
\alias{granularity,sim-method}
\alias{granularity<-,sim,numeric-method}
\title{Getter/Setter for Granularity}
\usage{
granularity(object)

granularity(object) <- value

\S4method{granularity}{sim}(object)

\S4method{granularity}{sim,numeric}(object) <- value
}
\arguments{
\item{object}{An S4 sim object.}

\item{value}{A numeric value that is beteen 0 and 100, right exclusive.}
}
\value{
A numeric vector representing granularity of simulation.
}
\description{
Granularity in terms of cores equals to \eqn{100 / Num_of_cores}, as each task can only be dispatched in mutiples of granularity. If CPU resources can (ideally) be dispatched in real numbers, granularity can take input of \code{0}.
}
