% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{is_well_performed}
\alias{is_well_performed}
\title{Check If Performacne is Good Enough}
\usage{
is_well_performed(score_result, target_score_1)
}
\arguments{
\item{score_result}{A sim_result object representing testing result on a testing batch.}

\item{target_score_1}{A numeric value representing target of score1 overall.}
}
\value{
A logical value \code{TRUE} if performance is good enough, \code{FALSE}, otherwise.
}
\description{
Check if the scheduling performance reaches target on score1.
}
\keyword{internal}
