% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_hoc_analysis.R
\name{label_performance_trace}
\alias{label_performance_trace}
\title{Label Traces Based on Their Performance}
\usage{
label_performance_trace(
  predict_info_quantiles,
  cut_off_prob,
  target = 1 - cut_off_prob,
  predict_info_statistics = NULL
)
}
\arguments{
\item{predict_info_quantiles}{A dataframe storing the prediction info.}

\item{cut_off_prob}{A numeric value representing cut off probability.}

\item{target}{A numeric value for the target of score1.}

\item{predict_info_statistics}{A dataframe storing statsitics of predictions. Can be \code{NULL}.}
}
\value{
A list containing update \code{predict_info_quantiles} and \code{predict_info_statistics}.
}
\description{
Label the traces based on their performance.
}
