% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{machine_survival}
\alias{machine_survival}
\title{Find Outcome of A Scheduled Foreground Job}
\usage{
machine_survival(machine_list, job_list, current_time, cores)
}
\arguments{
\item{machine_list}{A numeric vector containing actual values at current time.}

\item{job_list}{A dataframe containing information of all active jobs.}

\item{current_time}{A numeric value representing the current time.}

\item{cores}{A numeric value representing the number of threads for parallel programming for multiple traces, not supported for windows users.}
}
\value{
A list with keys representing decision and value job id falling into such decision.
}
\description{
Find Outcome of A Scheduled Foreground Job
}
\keyword{internal}
