% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{machine_update}
\alias{machine_update}
\title{Update Machine PI After Job Scheduling}
\usage{
machine_update(machine_pi_up, job_requestedCPU)
}
\arguments{
\item{machine_pi_up}{A numeric vector representing the prediction upper bound at different bins for a single machine.}

\item{job_requestedCPU}{A numeric number representing the requestedCPU for the job.}
}
\value{
A numeric vector representing the updated machine_pi_up after job scheduling.
}
\description{
Update Machine PI After Job Scheduling
}
\keyword{internal}
