% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microsoft_max_100}
\alias{microsoft_max_100}
\title{100 Samples of Azure Public Dataset V1}
\format{A data frame with 8000 rows and 100 columns:}
\source{
\url{https://github.com/Azure/AzurePublicDataset/blob/master/AzurePublicDatasetV1.md}
}
\usage{
microsoft_max_100
}
\description{
A dataframe conataining only maxes of traces from Microsoft Azure Data Center, with each column is one trace from one Virtual Machine.
The row name represents the time stamp of each obervation, the difference between observation is 300 seconds, the total length of each trace is almost 28 days.
The column name represents the hashed name of each virtual machine.
The value of each entry represents the maximum CPU value within the time granularity of 5 minutes.
}
\keyword{datasets}
