% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{normalize_predict_info}
\alias{normalize_predict_info}
\title{Normalize Parameter Info}
\usage{
normalize_predict_info(cut_off_prob, param_info)
}
\arguments{
\item{cut_off_prob}{A numeric vector representing cut off probability.}

\item{param_info}{A dataframe representing prediction information of different traces.}
}
\value{
A dataframe of normalized scoring information.
}
\description{
Change quantile information from column features to row features.
}
\keyword{internal}
