% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_hoc_analysis.R
\name{order_by_weighted_score1}
\alias{order_by_weighted_score1}
\title{Order the Performance of Traces Based on Weighted Score1}
\usage{
order_by_weighted_score1(
  predict_info_quantiles,
  cut_off_prob,
  target = 1 - cut_off_prob,
  adjustment = FALSE
)
}
\arguments{
\item{predict_info_quantiles}{A dataframe storing the prediction info.}

\item{cut_off_prob}{A numeric value representing the cut off probability to considered.}

\item{target}{A numeric value for the target of score1.}

\item{adjustment}{A logical value or \code{NA} representing whether to consider adjustment. \code{NA} represents only consider adjustment when score1 is below target.}
}
\value{
\code{predict_info_quantiles} ordered by weighted scores.
}
\description{
Order the performance of traces based on their weighted score1.
}
