% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{out_performs}
\alias{out_performs}
\title{Compare Performance of Two Models On A Test Batch}
\usage{
out_performs(score_result1, score_result2, target_score_1)
}
\arguments{
\item{score_result1}{A sim_result object representing testing result on a testing batch for model 1.}

\item{score_result2}{A sim_result object representing testing result on a testing batch for model 2.}

\item{target_score_1}{A numeric value representing target of score1 overall.}
}
\value{
A logical value \code{TRUE} if model 1 outperforms model 2, \code{FALSE} otherwise.
}
\description{
Check if model 1 outperforms model 2 on a test batch.
}
\keyword{internal}
