% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{param_setting1}
\alias{param_setting1}
\title{128 Numbers of Parameter Settings}
\format{
A dataframe with 128 rows and 10 columns:
Check \link{sim} for descriptions of the columns.
}
\usage{
param_setting1
}
\description{
A dataframe of parameter settings combining \code{window_size = c(12, 36)}, \code{cut_off_prob = c(0.005, 0.1)}, \code{granularity = c(100/32, 0)},
\code{train_size = c(2500, 5000)}, \code{model_num = c(1, 3)}, \code{update_freq = c(3, 6)}, \code{react_speed = list(c(1,1), c(1,3))}.
In order to pass into the \code{run_sim} argument, correct \code{name} column must be supplied.
}
\keyword{datasets}
