% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{parameter_parser}
\alias{parameter_parser}
\title{Convert Predicted Paraemeters into Lists}
\usage{
parameter_parser(predicted_param_df)
}
\arguments{
\item{predicted_param_df}{A dataframe of 1 row containing predicted parameters at specific timestamp.}
}
\description{
Convert Predicted Paraemeters into Lists
}
