% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot_ecdf_correlation}
\alias{plot_ecdf_correlation}
\title{Plot the ECDF of Correlation Or Cross-correlation of Different Types}
\usage{
plot_ecdf_correlation(
  dataset1,
  dataset2 = NULL,
  lags,
  freqs,
  corr_method = "pearson",
  response = c("max", "avg"),
  diffs = c(0, 0),
  diff_lags = c(1, 1),
  name,
  ...
)
}
\arguments{
\item{dataset1}{A numeric dataframe or matrix with each column as a trace of type 1.}

\item{dataset2}{A numeric dataframe or matrix with same dimension as \code{dataset1} representing a different type of trace.}

\item{lags}{A numeric vector representing the lag of autocorrelations or cross-correlations.}

\item{freqs}{A numeric vector representing the window size of observations to be aggregated.}

\item{corr_method}{A character representing the type of correlation to be calculated.}

\item{response}{A character vector of length two representing the type of aggregation, can be either \code{"max"} or \code{"avg"}}

\item{diffs}{A numeric vector of length two representing the order of differences on \code{dataset1} and \code{dataset2}.}

\item{diff_lags}{A numeric vector of length two representing which lag to use when differencing on \code{dataset1} and \code{dataset2}.}

\item{name}{A character that represents the identifier or name of the plot.}

\item{...}{Characters that represent the name of parent directories that will be passed to \code{write_location_check}.}
}
\description{
Plot correlation of \code{dataset1} with given lags or cross-correlation between \code{dataset1} and \code{dataset2} with given lags.
}
