% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot_ecdf_data}
\alias{plot_ecdf_data}
\title{Plot the ECDF of data}
\usage{
plot_ecdf_data(
  dataset,
  window_size,
  ecdf_or_hist = TRUE,
  mean_or_median = TRUE,
  response = "max",
  cores,
  name,
  ...
)
}
\arguments{
\item{dataset}{A numeric dataframe or matrix with each column as a trace.}

\item{window_size}{A numeric vector representing the window size of observations to be aggregated.}

\item{ecdf_or_hist}{A boolean representing the type of plot to be generated, \code{TRUE} means Emprical CDF, \code{FALSE} means histogram.}

\item{mean_or_median}{A boolean representing the type of aggregation across differenrt traces are used, \code{TRUE} means using mean, \code{FALSE} means using median.}

\item{response}{A character vector of length two representing the type of aggregation, can be either \code{"max"} or \code{"avg"}}

\item{cores}{A numeric value representing the number of threads for parallel programming, not supported for windows users.}

\item{name}{A character that represents the identifier or name of the plot.}

\item{...}{Characters that represent the name of parent directories that will be passed to \code{write_location_check}.}
}
\description{
Plot data points of \code{dataset} with given window sizes using ECDF or Histogram.
}
