% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot_ecdf_traces_performance}
\alias{plot_ecdf_traces_performance}
\title{Plot ECDF of Performance of Traces of A Control Variable}
\usage{
plot_ecdf_traces_performance(result_df, feature_name, adjusted, name, ...)
}
\arguments{
\item{result_df}{A dataframe containing the scores in different settings for the control variable parameter.}

\item{feature_name}{A character representing the name of the control variable whose value is stored in \code{result_df$feature} column.}

\item{adjusted}{A boolean representing whether the adjusted score will be plotted instead of original scores.}

\item{name}{A character that represents the identifier or name of the plot.}

\item{...}{Characters that represent the name of parent directories that will be passed to \code{write_location_check}.}
}
\description{
Plot ECDF of Performance of Traces of A Control Variable
}
