% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot_heatmap_correlations}
\alias{plot_heatmap_correlations}
\title{Plot the Heatmap of Correlation Or Cross-correlation of Different Window Sizes}
\usage{
plot_heatmap_correlations(
  dataset1,
  dataset2 = NULL,
  window_size1,
  window_size2 = NULL,
  response1,
  response2 = NULL,
  corr_method,
  cores,
  name,
  ...
)
}
\arguments{
\item{dataset1}{A numeric dataframe or matrix with each column as a trace of type 1.}

\item{dataset2}{A numeric dataframe or matrix with same dimension as \code{dataset1} representing a different type of trace.}

\item{window_size1}{A numeric vector representing the window size corresponding to \code{dataset1}.}

\item{window_size2}{A numeric vector representing the window size corresponding to \code{dataset2} or \code{dataset1} if \code{dataset2} is \code{NULL}.}

\item{response1}{A character vector representing the type of aggregation of \code{dataset1}, can be either \code{"max"} or \code{"avg"}}

\item{response2}{A character vector representing the type of aggregation of \code{dataset2}, can be either \code{"max"} or \code{"avg"} or \code{NULL} if \code{dataset2} is \code{NULL}.}

\item{corr_method}{A character representing the type of correlation to be calculated.}

\item{cores}{A numeric value representing the number of threads for parallel programming, not supported for windows users.}

\item{name}{A character that represents the identifier or name of the plot.}

\item{...}{Characters that represent the name of parent directories that will be passed to \code{write_location_check}.}
}
\description{
Plot correlation of \code{dataset1} or cross-correlation between \code{dataset1} and \code{dataset2}, in specified window sizes pairs.
}
