% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ar1_markov_sim.R, R/markov_sim.R,
%   R/sim_methods.R
\name{plot_sim}
\alias{plot_sim}
\alias{plot_sim,sim-method}
\title{Plot Simulation}
\usage{
plot_sim(object)

\S4method{plot_sim}{sim}(object)

\S4method{plot_sim}{sim}(object)

\S4method{plot_sim}{sim}(object)
}
\arguments{
\item{object}{An S4 sim objct.}
}
\value{
A plot object

A plot object

A plot object
}
\description{
Plot Simulation
}
