% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot_sim_charwise}
\alias{plot_sim_charwise}
\title{Plot Simulation Result Type Charwise}
\usage{
plot_sim_charwise(charwise_summ, name, ...)
}
\arguments{
\item{charwise_summ}{A dataframe containing the scores in all parameter settings and their performance.}

\item{name}{A character that represents the identifier or name of the plot.}

\item{...}{Characters that represent the name of parent directories that will be passed to \code{write_location_check}.}
}
\description{
Plot charwise result for simulation with each datapoint corresponds to average scores of all traces with one configuration.
}
