% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\name{plot_sim_overall}
\alias{plot_sim_overall}
\alias{plot_sim_overall,data.frame,character-method}
\title{Plot Simulation Result Type Overall}
\usage{
plot_sim_overall(overall_summ, result_loc)

\S4method{plot_sim_overall}{data.frame,character}(overall_summ, result_loc)
}
\arguments{
\item{overall_summ}{A dataframe containing the scores in all parameter settings and their performance.}

\item{result_loc}{A character that specify the path to which the result of simulations will be saved to.}
}
\value{
A plot object
}
\description{
Plot overall result for simulation with each datapoint corresponds to average scores of all traces with one configuration.
}
