% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot_sim_paramwise}
\alias{plot_sim_paramwise}
\title{Plot Simulation Result Type Tracewise}
\usage{
plot_sim_paramwise(param_score, target, name, ...)
}
\arguments{
\item{param_score}{A dataframe containing score information for all traces.}

\item{target}{A numeric value that is set to be the target of score 1 for all traces.}

\item{name}{A character that represents the identifier or name of the plot.}

\item{...}{Characters that represent the name of parent directories that will be passed to \code{write_location_check}.}
}
\description{
Plot tracewise result for simulation with each plot corresponds to the performance of one single trace.
}
