% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/arima_sim.R, R/sim_methods.R
\name{plot_sim_tracewise}
\alias{plot_sim_tracewise}
\alias{plot_sim_tracewise,arima_sim,character,list,data.frame,character-method}
\alias{plot_sim_tracewise,sim,character,list,data.frame,character-method}
\title{Plot Simulation Result Type Tracewise}
\usage{
plot_sim_tracewise(
  object,
  trace_name,
  trained_result,
  predict_info,
  result_loc
)

\S4method{plot_sim_tracewise}{arima_sim,character,list,data.frame,character}(
  object,
  trace_name,
  trained_result,
  predict_info,
  result_loc
)

\S4method{plot_sim_tracewise}{sim,character,list,data.frame,character}(
  object,
  trace_name,
  trained_result,
  predict_info,
  result_loc
)
}
\arguments{
\item{object}{An S4 sim object.}

\item{trace_name}{A character representing name of current trace.}

\item{trained_result}{A list or other class returend by \code{train_model}, containing trained model information.}

\item{predict_info}{A dataframe containing all the past predicted information.}

\item{result_loc}{A character that specify the path to which the result of simulations will be saved to.}
}
\description{
Plot tracewise result for simulation with each plot corresponds to the performance of one single trace.
}
