% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot_trace_removal_information}
\alias{plot_trace_removal_information}
\title{Plot the Diagnosis of Generated Traces}
\usage{
plot_trace_removal_information(
  score_change_info_list,
  cut_off_prob,
  target,
  window_size,
  granularity,
  adjustment,
  name,
  ...
)
}
\arguments{
\item{score_change_info_list}{A list with paste0(cut_off_prob, ",", target) as keys, and score change as values.}

\item{cut_off_prob}{A numeric vector that represents cut off probabilities.}

\item{target}{A numeric vector that represents the target of score 1.}

\item{window_size}{A numeric vector representing the window sizes to be displayed.}

\item{granularity}{A numeric value represneting granularity of response.}

\item{adjustment}{A logical value representing whether adjustment is accounted.}

\item{name}{A character representing identifier for the plot.}

\item{...}{Characters that represent the name of parent directories that will be passed to \code{write_location_check}.}
}
\description{
Plot the score change information for different combinations of different \code{cut_off_prob} and \code{window_size}.
}
