% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{pre_compute_models_foreground}
\alias{pre_compute_models_foreground}
\title{Precompute the Predicted Statistics of Foreground Models}
\usage{
pre_compute_models_foreground(
  load_foreground_result = NULL,
  param_setting_sim = NULL,
  additional_param_sim = list(),
  foreground_x = NULL,
  foreground_xreg = NULL,
  sim_length = 200,
  bins = c(1, 2, 6, 10, 14, 18, 22, 26, 30, 50, 80, 205),
  cores = parallel::detectCores()
)
}
\arguments{
\item{load_foreground_result}{\code{NULL} or a character representing the destined load/save file path.}

\item{param_setting_sim}{\code{NULL} or a dataframe representing the parameter setting of simulation, can be multiple roles corresponding to each \code{foreground_x} type.}

\item{additional_param_sim}{A list of additional parameter settings of simulation.}

\item{foreground_x}{\code{NULL} or a matrix or dataframe or a list of matrices or dataframes representing foreground observations.}

\item{foreground_xreg}{\code{NULL} or a matrix or dataframe or a list of matrices or dataframes or a list of lists of matrices or dataframes, representing the regressor corresponding to \code{foreground_x}.}

\item{sim_length}{\code{NULL} or a numeric integer representing the length of simulation. If exceeding the maximum length of each times series, the maximum length of each time series would be used.}

\item{bins}{\code{NULL} or a numeric vector representing the choices of window sizes to be kept track of.}

\item{cores}{\code{NULL} or a numeric integer representing the number of cores used for multiprocessing,}
}
\value{
A list containing all input variables and \code{"machine_bin_offs"} and \code{"fg_predict_info_lst"}.
}
\description{
Save/Load precomputed result in destined directory.
}
