% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_class.R
\docType{class}
\name{pred_result-class}
\alias{pred_result-class}
\alias{pred_result}
\title{An S4 Class to Represent A Prediction Result.}
\description{
An S4 Class to Represent A Prediction Result.
}
\section{Slots}{

\describe{
\item{\code{type}}{A character that represents the type of the prediction result, it should be either \code{"test"} for a testing batch, \code{"train"} for training model life time, \code{"trace"} for an entire trace, \code{"param"} for all traces in same parameter setting.}

\item{\code{score1.n}}{A numeric value that represents the performance on score 1 consistent with \code{type}.}

\item{\code{score1.w}}{A numeric value that represents the weight of \code{score1.n}.}
}}

