% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_sim.R
\name{predict_model_ar1}
\alias{predict_model_ar1}
\title{Schedule Jobs Using Predictions On Given Test Set}
\usage{
predict_model_ar1(
  last_obs,
  test_set,
  trained_result,
  window_size,
  cut_off_prob,
  granularity,
  schedule_policy,
  adjust_policy,
  mode
)
}
\arguments{
\item{last_obs}{The previous observation.}

\item{test_set}{The test set for scheduling and evaluations.}

\item{trained_result}{A list containing trained mean, coefficient, variance of residuals.}

\item{window_size}{The length of predictions.}

\item{cut_off_prob}{The level of uncertainty of prediction interval.}

\item{granularity}{The granularity of 100 percent of total cpu.}

\item{schedule_policy}{\code{"disjoint"} for scheduling at fixed time, \code{"dynamic"} for scheduling again immediately when failed.}

\item{adjust_policy}{\code{TRUE} for "backing off" strategy whenever a mistake is made.}

\item{mode}{\code{"max"} or \code{"avg"} which time series is used as \code{dataset}.}
}
\value{
A list containing the resulting scheduling informations.
}
\description{
Sequantially schedule jobs using predictions on provided test set.
}
