% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_markov_sim.R
\name{predict_model_ar1_markov}
\alias{predict_model_ar1_markov}
\title{Schedule Jobs Using Predictions On Given Test Set}
\usage{
predict_model_ar1_markov(
  test_set_max,
  test_set_avg,
  trained_result,
  window_size,
  cut_off_prob,
  granularity,
  schedule_policy,
  adjust_policy
)
}
\arguments{
\item{test_set_max}{The test set of max for scheduling and evaluations.}

\item{test_set_avg}{The test set of avg for scheduling and evaluations.}

\item{trained_result}{A list containing trained mean, coefficient, variance of residuals.}

\item{window_size}{The length of predictions.}

\item{cut_off_prob}{The level of uncertainty of prediction interval.}

\item{granularity}{The granularity of 100 percent of total cpu.}

\item{schedule_policy}{\code{"disjoint"} for scheduling at fixed time, \code{"dynamic"} for scheduling again immediately when failed.}

\item{adjust_policy}{\code{TRUE} for "backing off" strategy whenever a mistake is made.}
}
\value{
A list containing the resulting scheduling informations.
}
\description{
Sequantially schedule jobs using predictions on provided test set.
}
\keyword{internal}
