% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_pred.R
\name{predict_pred}
\alias{predict_pred}
\title{Predict Runtime of Jobs On Given Test Set}
\usage{
predict_pred(
  object,
  trained_result,
  test_x,
  test_xreg,
  predict_info,
  switch_status
)
}
\arguments{
\item{object}{A S4 pred object.}

\item{trained_result}{A trained object depending on the model used for training.}

\item{test_x}{A numeric vector representing the test set, containing maximum and average of actual observations.}

\item{test_xreg}{A dataframe representing the dataset that target dataset depends on for scheduling and evaluations.}

\item{predict_info}{A dataframe containing all the past predicted information.}

\item{switch_status}{A list containing all the information about current switches and identifiers.}
}
\value{
A dataframe containing the past predicted information and the current predicted information.
}
\description{
Sequantially predict running time of jobs on provided test set.
}
\keyword{internal}
