% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_pred.R
\name{predicting_pred}
\alias{predicting_pred}
\title{Simulation of Scheduling Jobs Based On Predictions}
\usage{
predicting_pred(object, x, xreg)
}
\arguments{
\item{object}{The corresponding trace/column in \code{dataset}.}

\item{x}{A numeric vector of length n representing the target dataset for predictions and evaluations.}

\item{xreg}{A matrix of length n by m representing the regressors.}
}
\value{
A list containing the resulting prediction informations.
}
\description{
Sequantially training and testing by scheduling jobs based on predictions on a single trace using AR1 Model.
}
\keyword{internal}
