% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predicting_sim_ar1}
\alias{predicting_sim_ar1}
\title{Simulation of Scheduling Jobs Based On Predictions With AR1 Model}
\usage{
predicting_sim_ar1(object, dataset_max, dataset_avg, cores, write_result)
}
\arguments{
\item{object}{A S4 sim object.}

\item{cores}{The number of threads for parallel programming for multiple traces, not supported for windows users.}

\item{write_result}{TRUE if the result of the experiment is written to a file.}

\item{testset_max}{A matrix of size \eqn{n \times m} representing the dataset of maximum for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}

\item{testset_avg}{A matrix of size \eqn{n \times m} representing the dataset of average for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}
}
\value{
A list containing the resulting prediction informations.
}
\description{
Sequantially training and testing by scheduling a job using AR1 Model.
}
\keyword{internal}
