% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var1_sim.R
\name{predicting_sim_var1}
\alias{predicting_sim_var1}
\title{Simulation of Scheduling Jobs Based On Predictions With VAR Model}
\usage{
predicting_sim_var1(
  param,
  dataset_max,
  dataset_avg,
  training_policy,
  schedule_policy,
  adjust_policy,
  cores,
  write_result
)
}
\arguments{
\item{param}{A vector containing necessary informations or hyperparameters for VAR model.}

\item{dataset_max}{A \eqn{n \times m} matrix, with each column is the time series of maxes of CPU information on a machine.}

\item{dataset_avg}{A \eqn{n \times m} matrix, with each column is the time series of avges of CPU information on a machine.#' @param cpu_required A vector of length \eqn{m}, each element is the size of the job trying to be scheduled on corresponding machine.}

\item{training_policy}{\code{"once"} for offline training, \code{"fixed"} for training at fixed time, \code{"dynamic"} for training when previous performance is bad.}

\item{schedule_policy}{\code{"disjoint"} for scheduling at fixed time, \code{"dynamic"} for scheduling again immediately when failed.}

\item{adjust_policy}{\code{TRUE} for "backing off" strategy whenever a mistake is made.}

\item{cores}{The number of threads for parallel programming for multiple traces, not supported for windows users.}

\item{write_result}{TRUE if the result of the experiment is written to a file.}
}
\value{
A dataframe containing the resulting scheduling informations.
}
\description{
Sequantially training and testing by scheduling a job using VAR Model.
}
\keyword{internal}
