% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_sim.R
\name{prediction_including_outlier_effect}
\alias{prediction_including_outlier_effect}
\title{Do Prediction Considering Outlier Occurrence Probability and Effect}
\usage{
prediction_including_outlier_effect(
  object,
  trained_result,
  pi_up,
  expected,
  level
)
}
\arguments{
\item{object}{A arima_sim object.}

\item{trained_result}{A tso object.}

\item{pi_up}{A dataframe containing predictions upper bound using normal distribution without considering outlier effect.}

\item{expected}{A dataframe containing predictions upper bound using normal distribution without considering outlier effect.}

\item{level}{A numeric vector representing the level of prediction interval.}
}
\value{
A list containing computed prediction upper bound, expected value and predicted parameters.
}
\description{
Do Prediction Considering Outlier Occurrence Probability and Effect
}
\keyword{internal}
