% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{replenish_jobs}
\alias{replenish_jobs}
\title{Replenish the Background Queue by Using Bg_predict_info}
\usage{
replenish_jobs(bg_predict_info, arrival_jobs, pin, num_jobs, emp_job_idx)
}
\arguments{
\item{bg_predict_info}{A dataframe to be regenerated from.}

\item{arrival_jobs}{A dataframe to be replenished.}

\item{pin}{A numeric value representing the current position to pick from bg_predict_info.}

\item{num_jobs}{A numeric value representing the resulting rows of arrival_jobs.}

\item{emp_job_idx}{A numeric value from which the empirical job indices are assigned.}
}
\value{
A numeric vector representing the updated machine_pi_up after job scheduling.
}
\description{
Replenish the Background Queue by Using Bg_predict_info
}
\keyword{internal}
