% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{res_dist}
\alias{res_dist}
\alias{res_dist<-}
\title{Getter/Setter for Residual Distribution}
\usage{
res_dist(object)

res_dist(object) <- value
}
\arguments{
\item{object}{An S4 sim object.}

\item{value}{A character value that can either be "norm" or "skew_norm".}
}
\description{
Res Dist is the Residual Distribution for ARIMA or VAR models.
}
