% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\name{response}
\alias{response}
\alias{response<-}
\title{Getter/Setter for Response}
\usage{
response(object)

response(object) <- value

\S4method{response}{sim}(object)

\S4method{response}{sim,numeric}(object) <- value
}
\arguments{
\item{object}{An S4 sim object}

\item{value}{A character value that is either \code{"max"} or \code{"avg"}.}
}
\description{
Response describes the type of response variable, the maximum of trace to predict or the average of trace to predict.
}
