% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{round_to_nearest}
\alias{round_to_nearest}
\title{Round To Nearest.}
\usage{
round_to_nearest(data, divisor, lower)
}
\arguments{
\item{data}{A numeic vector}

\item{divisor}{A numeric vector}

\item{lower}{If \code{TRUE}, then round down, if \code{FALSE}, then round up.}
}
\value{
The rounded data.
}
\description{
Round the \code{data} to the nearest number that can be divisible by \code{divisor}.
}
\keyword{internal}
