% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_sim}
\alias{run_sim}
\title{Run Simulation}
\usage{
run_sim(
  epoch_setting,
  x,
  xreg,
  result_loc = getwd(),
  cores = parallel::detectCores(),
  write_result = TRUE,
  plot_type
)
}
\arguments{
\item{epoch_setting}{A dataframe representing a specific parameter setting.}

\item{x}{A matrix of size n by m representing the target dataset for scheduling and evaluations.}

\item{xreg}{A matrix of length n by m representing the dataset that target dataset depends on for scheduling and evaluations.}

\item{result_loc}{A character that specify the path to which the result of simulations will be saved to. Default is your work directory.}

\item{cores}{A numeric numeb representing the number of threads for parallel programming for multiple traces, not supported for windows users.}

\item{write_result}{A logical value representing whether the final result of the simulation will be written as a csv file.}

\item{plot_type}{A character that can be one of "overall", "tracewise", "paramwise" or "none".}
}
\value{
An S4 sim result object.
}
\description{
The simulation dependes on the \code{type} attribute of the sim object, if \code{"scheduling"} is provided, simulation sequantially training and testing by scheduling a job with job size supplied by \code{cpu_required}, if \code{"predicting"} is supplied, sequential training and testing is made by predictions based on previous observations.
}
