% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{sample_moment_lag}
\alias{sample_moment_lag}
\title{Calculate Sample Covariance in Time Series}
\usage{
sample_moment_lag(dataset, k, r, s)
}
\arguments{
\item{dataset}{An observed dataset.}

\item{k}{lag}

\item{r}{power}
}
\value{
A numeric value calculated
}
\description{
This function calculates \eqn{\frac{1}{n} \sum_{i=1}^{n-k}(x_i - \bar{x})^r(x_{i+k} - \bar{x})^s}, a numeric estimator for \eqn{E[(X_t - E[X_t])^r(X_{t+k} - E[X_{t+k}])^s]}
}
\keyword{internal}
