% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{schedule_foreground}
\alias{schedule_foreground}
\title{Schedule A Job On Given Test Set}
\usage{
schedule_foreground(object_process, testset_max, testset_avg, cpu_required)
}
\arguments{
\item{object_process}{Hidden S4 sim object after training.}

\item{testset_max}{A matrix of size \eqn{n \times m} representing the test set of maximum for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}

\item{testset_avg}{A matrix of size \eqn{n \times m} representing the test set of average for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}

\item{cpu_required}{The size of the foreground job.}
}
\value{
A list containing the resulting scheduling informations.
}
\description{
Sequantially schedule a given job on given test set.
}
\keyword{internal}
