% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_logistic_sim.R
\name{schedule_foreground_ar1_logistic}
\alias{schedule_foreground_ar1_logistic}
\title{Schedule A Job On Given Test Set}
\usage{
schedule_foreground_ar1_logistic(
  test_set_max,
  test_set_avg,
  trained_result,
  window_size,
  cut_off_prob,
  cpu_required,
  granularity,
  schedule_policy,
  adjust_policy
)
}
\arguments{
\item{trained_result}{A list containing trained mean, coefficient, variance of residuals.}

\item{window_size}{The length of predictions.}

\item{cut_off_prob}{The maximum probability allowed to have next scheduling failing.}

\item{cpu_required}{A vector of length \eqn{m}, each element is the size of the job trying to be scheduled on corresponding machine.}

\item{granularity}{The granularity of 100 percent of total cpu.}

\item{schedule_policy}{\code{"disjoint"} for scheduling at fixed time, \code{"dynamic"} for scheduling again immediately when failed.}

\item{adjust_policy}{\code{TRUE} for "backing off" strategy whenever a mistake is made.}

\item{test_set}{The test set for scheduling and evaluations, the initial amount of observations that equals to window size are from training set.}
}
\value{
A list containing the resulting scheduling informations.
}
\description{
Sequantially schedule a given job on given test set.
}
\keyword{internal}
