% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\name{schedule_policy}
\alias{schedule_policy}
\alias{schedule_policy<-}
\alias{schedule_policy,sim-method}
\alias{schedule_policy<-,sim,character-method}
\title{Getter/Setter for Schedule Policy}
\usage{
schedule_policy(object)

schedule_policy(object) <- value

\S4method{schedule_policy}{sim}(object)

\S4method{schedule_policy}{sim,character}(object) <- value
}
\arguments{
\item{object}{An S4 sim object.}

\item{value}{A character value that can either be \code{"disjoint"} for scheduling with fixed frequency, or \code{"dynamic"} for scheduling with dynamic strategy.}
}
\value{
A character vector representing schedule policy of simulation.
}
\description{
Scheduling policy specifies the scheduling scheme, as in scheduling with fixed frequency, or retrying to schedule immediately after a job has failed.
}
