% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scheduling_sim}
\alias{scheduling_sim}
\title{Simulation of Scheduling A Job}
\usage{
scheduling_sim(
  object,
  dataset_max,
  dataset_avg,
  cpu_required,
  cores,
  write_result,
  plot_type
)
}
\arguments{
\item{object}{An S4 sim object.}

\item{dataset_max}{A matrix of size \eqn{n \times m} representing the dataset of maximum for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}

\item{dataset_avg}{A matrix of size \eqn{n \times m} representing the dataset of average for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}

\item{cpu_required}{A vector of length \eqn{m}, each element is the size of the job trying to be scheduled on corresponding machine.}

\item{cores}{The number of threads for parallel programming for multiple traces, not supported for windows users.}

\item{write_result}{TRUE if the result of the experiment is written to a file.}

\item{plot_type}{A character that can be one of "overall", "tracewise", "paramwise" or "none".}
}
\value{
A corresponding S4 sim result object.
}
\description{
Sequantially training and testing by scheduling a job.
}
\keyword{internal}
