% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_logistic_sim.R
\name{scheduling_sim_ar1_logistic}
\alias{scheduling_sim_ar1_logistic}
\title{Simulation of Scheduling A Job With AR1-Logistic Model}
\usage{
scheduling_sim_ar1_logistic(
  param,
  dataset_avg,
  dataset_max,
  cpu_required,
  training_policy,
  schedule_policy,
  adjust_policy,
  cores,
  write_result
)
}
\arguments{
\item{param}{A vector containing necessary informations or hyperparameters for AR1-Logistic model.}

\item{cpu_required}{A vector of length \eqn{m}, each element is the size of the job trying to be scheduled on corresponding machine.}

\item{training_policy}{\code{"once"} for offline training, \code{"fixed"} for training at fixed time, \code{"dynamic"} for training when previous performance is bad.}

\item{schedule_policy}{\code{"disjoint"} for scheduling at fixed time, \code{"dynamic"} for scheduling again immediately when failed.}

\item{adjust_policy}{\code{TRUE} for "backing off" strategy whenever a mistake is made.}

\item{cores}{The number of threads for parallel programming for multiple traces, not supported for windows users.}

\item{write_result}{TRUE if the result of the experiment is written to a file.}

\item{dataset}{A \eqn{n \times m} matrix, with each column is the time series of maxes and avges of CPU information on a machine.}

\item{mode}{\code{"max"} or \code{"avg"} which time series is used as \code{dataset}.}
}
\value{
A dataframe containing the resulting scheduling informations.
}
\description{
Sequantially training and testing by scheduling a job using AR1-Logistic Model.
}
\keyword{internal}
