% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_arguments.R
\name{set_parameters}
\alias{set_parameters}
\title{Set Parameters}
\usage{
set_parameters(param_name, value)
}
\arguments{
\item{param_name}{A string that is either \code{window_size}, \code{cut_off_prob}, \code{granularity}, \code{train_size}, \code{update_freq} or \code{NULL}}

\item{value}{A numeric vector that needs to be compactable with the \code{param_name} provided.}
}
\description{
Set the value of a particular parameter.
}
\examples{
set_parameters("window_size", c(12, 36))
}
