% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_arguments.R
\name{set_result_location}
\alias{set_result_location}
\title{Change Result Location.}
\usage{
set_result_location(path = NULL)
}
\arguments{
\item{path}{The path you wish to store the result to or \code{NULL}.}
}
\description{
Set the location to store the result of the simulations to given \code{path} or the default work directory if \code{NULL} is provided.
}
\examples{
set_result_location()
}
