% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{show_result}
\alias{show_result}
\title{Print The Result Of Simulation}
\usage{
show_result(param_predict_info, show_msg = TRUE)
}
\arguments{
\item{param_predict_info}{A S4 sim result object.}

\item{show_msg}{A logical value representing whether messages will printed.}
}
\value{
A list consists of summary of simulation result.
}
\description{
Show the result of a simulation for one specific parameter setting.
}
\keyword{internal}
