% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_class.R, R/ar1_markov_sim.R, R/ar1_sim.R,
%   R/markov_sim.R, R/var1_sim.R
\docType{class}
\name{sim-class}
\alias{sim-class}
\alias{sim}
\alias{ar1_markov_sim-class}
\alias{ar1_markov_sim}
\alias{ar1_sim-class}
\alias{ar1_sim}
\alias{markov_sim-class}
\alias{markov_sim}
\alias{var1_sim-class}
\alias{var1_sim}
\title{An S4 Class to Represent A Simulation.}
\arguments{
\item{state_num}{A numeric number that represents the number of states in Markov chain.}

\item{res_dist}{The distribution of residual.}

\item{reg_num}{The number of past regressive observations needed to forecast next observation.}
}
\description{
An S4 Class to Represent A Simulation.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character that represents the name of the simulation.}

\item{\code{type}}{A character that specify the type of simulation to run, this can either be \code{"scheduling"} or \code{"predicting"}.}

\item{\code{window_size}}{A numeric vector that can only be integers to specify how many observations to be aggregated as one.}

\item{\code{cut_off_prob}}{A numeric vector that is the the maximum probability allowed to have next scheduling failing if \code{type = "scheduling"}, and the level of the prediction interval if \code{type = "predicting"}. Default values are \code{0.005, 0.01, 0.1}.}

\item{\code{granularity}}{A numeric vector that specify the amount of CPU usage can be scheduled by one core, if \code{0} is provided, then granularity is not considered. Default values are \code{100/32, 100/64, 0}.}

\item{\code{train_size}}{A numeric vector that specify the training size used for simulations. Default values are \code{2000, 3000, 4000}.}

\item{\code{update_freq}}{A numeric vector that specify the length of testing after each training step, also the amount of step to update after testing step is complete. Default values are \code{2000, 3000, 4000}.}

\item{\code{train_policy}}{A character that specify the training policy, this can either be \code{"once"}, \code{"fixed"} and \code{dynamic}. Default value is \code{"fixed"}.}

\item{\code{schedule_policy}}{A character that specify the scheduling policy, this can either be \code{"disjoint"} or \code{"dynamic"}. Default value is \code{"dynamic"}.}

\item{\code{adjust_policy}}{A character that specify the adjustment policy, this can either be \code{"back_off"} or \code{"none"}. Defaut is \code{"none"}.}

\item{\code{tolerance1}}{A numeric vector that specify the minimum quantile of past performance needs to be achieved for score1, otherwise, re-train signal will be sent.}

\item{\code{tolerance2}}{A numeric vector that specify the minimum quantile of past performance needs to be achieved for score2, otherwise, re-train signal will be sent.}

\item{\code{response}}{A character that specify the targeting trace to be tested on, this can either be \code{"max"} or \code{"avg"} for max traces and average traces respectively.}

\item{\code{result_loc}}{A character that specify the path to which the result of simulations will be saved to. Default is your work directory.}
}}

