% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_class.R
\name{sim_result-class}
\alias{sim_result-class}
\title{An S4 Class to Represent Result of Simulation}
\description{
This Class is not created mannually by user, but produced as the result of the main simulation functions: \code{run_sim}.
}
\section{Slots}{

\describe{
\item{\code{result}}{A dataframe object storing the result of simulation for each trace, each row represents one trace, and the columns represent the performance information.}

\item{\code{summ}}{A list representing the summary of all traces, with four elements: average, aggregated score 1 and score 2.}
}}

\seealso{
\code{\link{sim-class}} for super class.
}
