% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{skew_norm_param_estimation}
\alias{skew_norm_param_estimation}
\title{Estimate Parameters of Skew Normal Distribution Using MOM}
\usage{
skew_norm_param_estimation(res)
}
\arguments{
\item{res}{A numeric vector of residuals of fitted model.}
}
\value{
A list containing the xi, omega and alpha.
}
\description{
Estimate Parameters of Skew Normal Distribution Using MOM
}
\keyword{internal}
