% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{skew_norm_param_prediction}
\alias{skew_norm_param_prediction}
\title{Predict Parameters of Skew Normal Distribution Using MOM}
\usage{
skew_norm_param_prediction(object, trained_result, predicted_mean, level)
}
\arguments{
\item{object}{A numeric vector of residuals of fitted model.}

\item{trained_result}{A tso or Arima object containing trained parameters.}

\item{predicted_mean}{A numeric vector representing predicted mean under normal distribution assumption.}

\item{level}{A numeric vector representing the confidence level.}
}
\value{
A list containing the xi, omega and alpha.
}
\description{
Predict Parameters of Skew Normal Distribution Using MOM
}
\keyword{internal}
