% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{split_to_uni}
\alias{split_to_uni}
\title{Split A Sim Object Into Sim Objects With Length 1 Slots}
\usage{
split_to_uni(object)
}
\arguments{
\item{object}{An S4 sim object}
}
\value{
A list containing all the sim objects with uni-length slots
}
\description{
Split A Sim Object Into Sim Objects With Length 1 Slots
}
\keyword{internal}
