% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{state_num}
\alias{state_num}
\alias{state_num<-}
\title{Getter/Setter for Number of States}
\usage{
state_num(object)

state_num(object) <- value
}
\arguments{
\item{object}{An S4 sim object.}

\item{value}{A numeric value that is a positive integer.}
}
\description{
State num is the number of states for Markov Models or Mixed models that contains Markov chain components.
}
