% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{svt_combine_simulation}
\alias{svt_combine_simulation}
\title{Single Combined  Simulation with Synthetic Background Jobs}
\usage{
svt_combine_simulation(
  ts_num,
  sim_object,
  window_multiplier,
  sim_length,
  bin,
  machine_bin_offs,
  fg_predict_info_lst,
  foreground_x,
  foreground_xreg,
  write_type,
  ...
)
}
\arguments{
\item{ts_num}{A numeric value representing the index of current trace.}

\item{sim_object}{A sim object representing foreground setting.}

\item{window_multiplier}{A numeric number representing the multiplier for delayed time and scheduled time.}

\item{sim_length}{A numeric integer representing the length of time for simulation, training size excluded.}

\item{bin}{A numeric integer representing a specific partioning of time.}

\item{machine_bin_offs}{A dataframe containing the offset and bin information for each element of the list.}

\item{foreground_x}{A dataframe representing foreground dataset.}

\item{foreground_xreg}{\code{NULL} or a dataframe or a list of dataframes representing the regressor of foreground data.}

\item{write_type}{A character representing whether to record the simulated information.}

\item{...}{Additional arguments passed into \code{write_sim_result}.}

\item{fg_predict_info_offs}{A list containing foreground simulation predictions at each time.}
}
\value{
A dataframe representing the score for this single trace.
}
\description{
Single Combined  Simulation with Synthetic Background Jobs
}
\keyword{internal}
