% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{svt_machine_survival}
\alias{svt_machine_survival}
\title{Find Outcome of A Scheduled Foreground Job for Single Machine}
\usage{
svt_machine_survival(
  scheduled_jobs,
  machine_available_info,
  current_time,
  window_multiplier,
  algo = "greedy",
  max_combn = 5000
)
}
\arguments{
\item{scheduled_jobs}{A dataframe containing information of all active jobs.}

\item{machine_available_info}{A numeric value representing the current time.}

\item{current_time}{A numeric value representing the current time.}

\item{window_multiplier}{A numeric number representing the multiplier for delayed time and scheduled time.}

\item{algo}{A character value representing the algorithm of choice. Default is \code{"greedy"} which is fast but does not yield the best choice. \code{"dynamic"} is slower but ultimate produces the best selections.}

\item{max_comb}{A numeric value representing the maximum number of combination in finding optimal combinations of jobs, used for limit memory use, only used when \code{algo = "dynmaic"}. Default value is \code{5000}.}
}
\value{
A numeric vector containing job ids to be removed.
}
\description{
Find Outcome of A Scheduled Foreground Job for Single Machine
}
\keyword{internal}
