% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{svt_predicting_sim}
\alias{svt_predicting_sim}
\title{Simulation of Scheduling Jobs Based On Predictions On A Single Trace.}
\usage{
svt_predicting_sim(
  ts_num,
  object,
  x,
  xreg = NULL,
  start_point = 1,
  wait_time = 0,
  write_type,
  plot_type,
  ...
)
}
\arguments{
\item{ts_num}{The corresponding trace/column in \code{dataset}.}

\item{x}{A numeric vector of length n representing the target dataset for scheduling and evaluations.}

\item{xreg}{A numeric vector of length n representing the external regressor.}

\item{start_point}{A numeric number that represents the starting point of the simulation. Default value is \code{1}.}

\item{wait_time}{A numeric number that represents the time between training and testing. Default value is \code{0}.}

\item{write_type}{A character that represents how to write the result of simulation, can be one of "charwise", "tracewise", "paramwise" or "none".}

\item{plot_type}{A character that can be one of "charwise", "tracewise", "paramwise" or "none".}

\item{...}{Characters that represent the name of parent directories that will be passed to \code{write_location_check}.}
}
\value{
A list containing the resulting prediction informations.
}
\description{
Sequantially training and testing by scheduling jobs based on predictions on a single trace.
}
\keyword{internal}
