% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{svt_predicting_sim_ar1}
\alias{svt_predicting_sim_ar1}
\title{Simulation of Scheduling Jobs Based On Predictions On A Single Trace With AR1 Model}
\usage{
svt_predicting_sim_ar1(ts_num, object, dataset_max, dataset_avg)
}
\arguments{
\item{ts_num}{The corresponding trace/column in \code{dataset}.}

\item{testset_max}{A matrix of size \eqn{n \times m} representing the dataset of maximum for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}

\item{testset_avg}{A matrix of size \eqn{n \times m} representing the dataset of average for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}
}
\value{
A list containing the resulting prediction informations.
}
\description{
Sequantially training and testing by scheduling jobs based on predictions on a single trace using AR1 Model.
}
\keyword{internal}
