% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{svt_scheduleing_sim}
\alias{svt_scheduleing_sim}
\title{Simulation of Scheduling A Job On A Single Trace}
\usage{
svt_scheduleing_sim(ts_num, object, dataset_max, dataset_avg, cpu_required)
}
\arguments{
\item{ts_num}{The corresponding trace/column in \code{dataset_max} or \code{dataset_avg}.}

\item{object}{An S4 sim object.}

\item{dataset_max}{A matrix of size \eqn{n \times m} representing the dataset of maximum for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}

\item{dataset_avg}{A matrix of size \eqn{n \times m} representing the dataset of average for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}

\item{cpu_required}{A vector of length \eqn{m}, each element is the size of the job trying to be scheduled on corresponding machine.}
}
\value{
A list containing the resulting scheduling informations.
}
\description{
Sequantially training and testing by schedulingh a job on a single trace.
}
\keyword{internal}
