% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_logistic_sim.R
\name{svt_scheduleing_sim_ar1_logistic}
\alias{svt_scheduleing_sim_ar1_logistic}
\title{Simulation of Scheduling A Job On A Single Trace With AR1-Logistic Model}
\usage{
svt_scheduleing_sim_ar1_logistic(
  ts_num,
  dataset_avg,
  dataset_max,
  cpu_required,
  train_size,
  window_size,
  update_freq,
  cut_off_prob,
  granularity,
  training_policy,
  tolerance,
  schedule_policy,
  adjust_policy
)
}
\arguments{
\item{ts_num}{The corresponding trace/column in \code{dataset}.}

\item{cpu_required}{A vector of length \eqn{m}, each element is the size of the job trying to be scheduled on corresponding machine.}

\item{train_size}{The length of data used for training.}

\item{window_size}{The length of predictions.}

\item{update_freq}{The length of testing on scheduing decision each iteration.}

\item{cut_off_prob}{The maximum probability allowed to have next scheduling failing.}

\item{granularity}{The granularity of 100 percent of total cpu.}

\item{training_policy}{\code{"once"} for offline training, \code{"fixed"} for training at fixed time, \code{"dynamic"} for training when previous performance is bad.}

\item{tolerance}{The tolerance level of retrain, the quantile of previous performance.}

\item{schedule_policy}{\code{"disjoint"} for scheduling at fixed time, \code{"dynamic"} for scheduling again immediately when failed.}

\item{adjust_policy}{\code{TRUE} for "backing off" strategy whenever a mistake is made.}

\item{dataset}{A \eqn{n \times m} matrix, with each column is the time series of maxes and avges of CPU information on a machine.}
}
\value{
A list containing the resulting scheduling informations.
}
\description{
Sequantially training and testing by schedulingh a job on a single trace using AR1-Logistic Model.
}
\keyword{internal}
