% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{trace_reconstruct}
\alias{trace_reconstruct}
\title{Reconstruct More Granular Trace Based on Average Trace.}
\usage{
trace_reconstruct(
  avg_trace,
  max_trace = NULL,
  orig_rate,
  new_rate,
  h = new_rate,
  d,
  tune_factor = seq(from = 0.08, to = 2, by = 0.01)
)
}
\arguments{
\item{avg_trace}{A numeric vector that represents the average trace taken in a fixed sampled rate.}

\item{max_trace}{A numeric vector that represents the maximum trace with sample length as \code{max_trace} and same sampled rate. Used for fine tuning of the generated trace, or \code{NULL}. Default value is \code{NULL}.}

\item{orig_rate}{A numeric positive integer that is typicall smaller that the frequency of \code{avg_trace} and \code{max_trace}.}

\item{new_rate}{A numeric positive integer that is typicall smaller that the frequency of \code{avg_trace} and \code{max_trace}.}

\item{h}{A numeric value representing the granularity for numerical differentiation, \eqn{(f(x+h) - f(x)) / h}, must be smaller or equal to the value of \code{new_freq}. Default value is \code{new_rate}.}

\item{d}{A numeric integer representing the distance of nearest neighbourhood to take into account in sinc function. Passed into function \code{signal::resample}.}

\item{tune_factor}{A numeric vector representing the percentage change in fine tuning that should be taken into account.}
}
\description{
Reconstruct a more granular trace of frequency \code{new_sample_freq} based on Nyquist–Shannon sampling theorem with assumption that the frequency of the original trace must not contain frequencies higher than 1/2 of the \code{new_sample_freq}.
}
