% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_sim.R
\name{train_ar1}
\alias{train_ar1}
\title{Train AR1 Model.}
\usage{
train_ar1(train_dataset)
}
\arguments{
\item{train_dataset}{A vector of numeric value.}
}
\value{
A list consisting trained coefficients, mean, and variance of residuals.
}
\description{
Train AR1 model using training set provided.
}
