% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_logistic_sim.R
\name{train_ar1_logistic}
\alias{train_ar1_logistic}
\title{Train AR1_logistic Model}
\usage{
train_ar1_logistic(train_dataset_max, train_dataset_avg, cpu_required)
}
\arguments{
\item{train_dataset_max}{A vector of numeric value of max.}

\item{train_dataset_avg}{A vector of numeric value of avg}
}
\value{
A list consisting trained coefficients, mean, and variance of residuals.
}
\description{
Train ar1 and logistic models using training set provided.
}
\keyword{internal}
