% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_markov_sim.R
\name{train_ar1_markov}
\alias{train_ar1_markov}
\title{Train Markov Model}
\usage{
train_ar1_markov(
  disjoint_dataset_avg,
  overlapping_dataset_max,
  overlapping_dataset_avg,
  state_num
)
}
\arguments{
\item{disjoint_dataset_avg}{A vector of training set average aggregated using disjoint schemes.}

\item{overlapping_dataset_max}{A vector of training set average aggregated using overlapping schemes.}

\item{overlapping_dataset_avg}{A vector of training set max aggregated using overlapping schemes.}

\item{state_num}{number of states for the Markov Chain.}
}
\value{
A transitional matrix for the Markov Chain.
}
\description{
Train Markov model using training set provided.
}
\keyword{internal}
