% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_sim.R
\name{train_markov}
\alias{train_markov}
\title{Train Markov Model}
\usage{
train_markov(dataset, state_num)
}
\arguments{
\item{dataset}{A vector of numeric value.}

\item{state_num}{number of states for the Markov Chain}
}
\value{
A transitional matrix for the Markov Chain
}
\description{
Train Markov model using training set provided.
}
