% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/anovatree_pred.R, R/arima_sim.R,
%   R/auto_arima_sim.R, R/gmm_pred.R, R/markov_sim.R, R/surtree_pred.R
\name{train_model}
\alias{train_model}
\alias{train_model,anovatree_pred,numeric,numeric-method}
\alias{train_model,arima_sim,numeric,numeric-method}
\alias{train_model,auto_arima_sim,numeric,numeric-method}
\alias{train_model,gmm_pred,numeric,numeric-method}
\alias{train_model,markov_sim,numeric,numeric-method}
\alias{train_model,surtree_pred,numeric,numeric-method}
\title{Train Model}
\usage{
train_model(object, train_x, train_xreg)

\S4method{train_model}{anovatree_pred,numeric,numeric}(object, train_x, train_xreg)

\S4method{train_model}{arima_sim,numeric,numeric}(object, train_x, train_xreg)

\S4method{train_model}{auto_arima_sim,numeric,numeric}(object, train_x, train_xreg)

\S4method{train_model}{gmm_pred,numeric,numeric}(object, train_x, train_xreg)

\S4method{train_model}{markov_sim,numeric,numeric}(object, train_x, train_xreg)

\S4method{train_model}{surtree_pred,numeric,numeric}(object, train_x, train_xreg)
}
\arguments{
\item{object}{An S4 sim object.}

\item{train_x}{A numeric of length m representing the training set.}

\item{train_xreg}{A numeric or matrix of length or row number m representing the additional regressors for training.}
}
\value{
A list containing trained result.
}
\description{
This is a generic function that trains model according to the input object type, with additional arguments supplied by attributes of the object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = anovatree_pred,train_x = numeric,train_xreg = numeric}: Train ARMA Model specific to anovatree_pred object.

\item \code{object = arima_sim,train_x = numeric,train_xreg = numeric}: Train ARMA Model specific to arima_sim object.

\item \code{object = auto_arima_sim,train_x = numeric,train_xreg = numeric}: Train ARMA Model specific to arima_sim object.

\item \code{object = gmm_pred,train_x = numeric,train_xreg = numeric}: Train ARMA Model specific to gmm_pred object.

\item \code{object = markov_sim,train_x = numeric,train_xreg = numeric}: Train Markov Model specific to markov_sim object.

\item \code{object = surtree_pred,train_x = numeric,train_xreg = numeric}: Train ARMA Model specific to surtree_pred object.
}}

