% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.sim.R
\name{train_mvt_model}
\alias{train_mvt_model}
\title{Train VAR Model}
\usage{
train_mvt_model(train_dataset_max, train_dataset_avg, p, q)
}
\arguments{
\item{train_dataset}{Two vectors of numeric value: one trace for max and one trace for avg}
}
\value{
A list consisting trained coefficients, standard errors, and residuals.
}
\description{
Train VAR model using training set provided.
}
