% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\docType{methods}
\name{train_policy}
\alias{train_policy}
\alias{train_policy<-}
\alias{train_policy,sim-method}
\alias{train_policy<-,sim,character-method}
\title{Getter/Setter for Training Policy}
\usage{
train_policy(object)

train_policy(object) <- value

\S4method{train_policy}{sim}(object)

\S4method{train_policy}{sim,character}(object) <- value
}
\arguments{
\item{object}{An S4 sim object.}

\item{value}{A character value that can either be \code{"once"} for offline training, or \code{"fixed"} for online training with fixed frequency, or \code{"dynamic"} for online training based on evalutions.}
}
\value{
A character representing training policy of simulation.
}
\description{
Training policy specifies the training scheme, as in offline training, online training with fixed frequency, or online training based on current and previous evaluations.
}
