% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\name{train_size}
\alias{train_size}
\alias{train_size<-}
\alias{train_size,sim-method}
\alias{train_size<-,sim,numeric-method}
\title{Getter/Setter for Traning Size}
\usage{
train_size(object)

train_size(object) <- value

\S4method{train_size}{sim}(object)

\S4method{train_size}{sim,numeric}(object) <- value
}
\arguments{
\item{object}{An S4 sim object}

\item{value}{A numeric value that is a positive integer.}
}
\value{
A numeric vector representing training size of simulation.
}
\description{
Traning size is the amount of observations used to train a model, both in offline training and online training.
}
