% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var1_sim.R
\name{train_var1}
\alias{train_var1}
\title{Train VAR Model}
\usage{
train_var1(train_dataset_max, train_dataset_avg)
}
\arguments{
\item{train_dataset_max}{A numeric vector representing max trace.}

\item{train_dataset_avg}{A numeric vector representing avg trace.}
}
\value{
A list consisting trained coefficients, standard errors, and residuals.
}
\description{
Train VAR model using training set provided.
}
\keyword{internal}
