% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\docType{methods}
\name{update_freq}
\alias{update_freq}
\alias{update_freq<-}
\alias{update_freq,sim-method}
\alias{update_freq<-,sim,numeric-method}
\title{Getter/Setter for Update Frequency}
\usage{
update_freq(object)

update_freq(object) <- value

\S4method{update_freq}{sim}(object)

\S4method{update_freq}{sim,numeric}(object) <- value
}
\arguments{
\item{object}{An S4 sim object.}

\item{value}{A numeric value that is a positive integer.}
}
\value{
A numeric vector representing update frequency of simulation.
}
\description{
Update frequency is the step size of each update in an online simulation with rolling windows, it also serves as the size of (batch) test set after each training step.
}
