% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\docType{methods}
\name{window_size}
\alias{window_size}
\alias{window_size<-}
\alias{window_size,sim-method}
\alias{window_size<-,sim,numeric-method}
\title{Getter/Setter for Window Size}
\usage{
window_size(object)

window_size(object) <- value

\S4method{window_size}{sim}(object)

\S4method{window_size}{sim,numeric}(object) <- value
}
\arguments{
\item{object}{An S4 sim object.}

\item{value}{A numeric value that is a postive integer.}
}
\value{
A numeric vector representing window size of simulation.
}
\description{
A window size is the length of aggregation of observations as an alternative or reduction of extrapolation in forecasting.
}
