% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{write_location_check}
\alias{write_location_check}
\title{Check Write Location}
\usage{
write_location_check(..., file_name)
}
\arguments{
\item{...}{The name of parent directories that will be used for path of parent directory.}

\item{file_name}{The name of file.}
}
\description{
Check Write Location
}
\keyword{internal}
