% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{write_sim_result}
\alias{write_sim_result}
\title{Write The Result Of Simulation}
\usage{
write_sim_result(overall_summ, result_loc)
}
\arguments{
\item{overall_summ}{A dataframe containing the scores in all parameter settings and their performance.}

\item{result_loc}{A character that specify the path to which the result of simulations will be saved to.}
}
\description{
Write the result of a simulation to csv file.
}
\keyword{internal}
