% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_score_trace}
\alias{check_score_trace}
\title{Check Score Of An Entire Trace.}
\usage{
check_score_trace(cut_off_prob, predict_info)
}
\arguments{
\item{cut_off_prob}{A numeric vector representing cut off probability.}

\item{predict_info}{A dataframe representing past prediction information.}
}
\value{
A sim_result object.
}
\description{
Check the score information of an entire trace based on actual observations and predictions.
}
\keyword{internal}
