% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ar1_markov.sim.R, R/ar1_sim.R,
%   R/markov_sim.R, R/var1_sim.R
\docType{methods}
\name{compute_pi_up}
\alias{compute_pi_up}
\alias{compute_pi_up,ar1_markov_sim_process-method}
\alias{compute_pi_up,ar1_sim_process-method}
\alias{compute_pi_up,markov_sim_process-method}
\alias{compute_pi_up,var1_sim_process-method}
\title{Compute Prediction Interval Upper Bound}
\usage{
compute_pi_up(object)

\S4method{compute_pi_up}{ar1_markov_sim_process}(object)

\S4method{compute_pi_up}{ar1_sim_process}(object)

\S4method{compute_pi_up}{markov_sim_process}(object)

\S4method{compute_pi_up}{var1_sim_process}(object)
}
\arguments{
\item{object}{An S4 sim process object with prediction information.}
}
\value{
A numeric value representing the maximum of prediction interval upper bound in the forecasting steps.
}
\description{
This is a generic function that computes the upper bound of prediction interval according to the type of input object.
}
\section{Methods (by class)}{
\itemize{
\item \code{ar1_markov_sim_process}: Compute prediction interval Upper Bound based on trained AR1-Markov Model.

\item \code{ar1_sim_process}: Compute prediction interval Upper Bound based on trained AR1 Model.

\item \code{markov_sim_process}: Compute prediction interval Upper Bound based on trained Markov Model.

\item \code{var1_sim_process}: Compute prediction interval Upper Bound based on trained AR1 Model.
}}

