% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{compute_pi_up}
\alias{compute_pi_up}
\title{Compute Prediction Interval Upper Bound For Forcasting.}
\usage{
compute_pi_up(mu, varcov, predict_size, cut_off_prob)
}
\arguments{
\item{mu}{The predicted mean of next observations.}

\item{varcov}{The variance covariance matrix of predictions of next observations.}

\item{predict_size}{The number of steps to predict forward.}

\item{cut_off_prob}{The level of uncertainty of prediction interval.}
}
\value{
The prediction upper bounds, if \code{predict_size} is greater than 1, a vector is returned.
}
\description{
Computer prediction using predicted value and variance, can be adjusted to granularity.
}
