% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{do_prediction}
\alias{do_prediction}
\title{Do Prediction}
\usage{
do_prediction(object, last_obs_max, last_obs_avg, predict_size, level)
}
\arguments{
\item{object}{An S4 sim process object with trained model information.}

\item{last_obs_max}{A numeric value representing the last observation of maximum.}

\item{last_obs_avg}{A numeric value representing the last observation of average.}

\item{predict_size}{The number of window sizes to predict forword.}

\item{level}{The level in \eqn{Pr(Y_{t+1}|Y_{t}) \leq level} to be computed.}
}
\value{
The same S4 process object with prediction information.
}
\description{
This is a generic function that do prediction according to the input object type.
}
