% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ar1_markov.sim.R, R/ar1_sim.R,
%   R/markov_sim.R, R/var1_sim.R
\docType{methods}
\name{do_prediction}
\alias{do_prediction}
\alias{do_prediction,ar1_markov_sim_process,numeric,numeric,numeric,numeric-method}
\alias{do_prediction,ar1_sim_process,numeric,numeric,numeric,numeric-method}
\alias{do_prediction,markov_sim_process,numeric,numeric,numeric,numeric-method}
\alias{do_prediction,var1_sim_process,numeric,numeric,numeric,numeric-method}
\title{Do Prediction}
\usage{
do_prediction(object, last_obs_max, last_obs_avg, predict_size, level)


  \S4method{do_prediction}{ar1_markov_sim_process,numeric,numeric,numeric,numeric}(object,
  last_obs_max, last_obs_avg, predict_size, level)


  \S4method{do_prediction}{ar1_sim_process,numeric,numeric,numeric,numeric}(object,
  last_obs_max, last_obs_avg, predict_size, level)


  \S4method{do_prediction}{markov_sim_process,numeric,numeric,numeric,numeric}(object,
  last_obs_max, last_obs_avg, predict_size, level)


  \S4method{do_prediction}{var1_sim_process,numeric,numeric,numeric,numeric}(object,
  last_obs_max, last_obs_avg, predict_size, level)
}
\arguments{
\item{object}{An S4 sim process object with trained model information.}

\item{last_obs_max}{A numeric value representing the last observation of maximum.}

\item{last_obs_avg}{A numeric value representing the last observation of average.}

\item{predict_size}{The number of window sizes to predict forword.}

\item{level}{The level in \eqn{Pr(Y_{t+1}|Y_{t}) \leq level} to be computed.}
}
\value{
The same S4 process object with prediction information.
}
\description{
This is a generic function that do prediction according to the input object type.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = ar1_markov_sim_process,last_obs_max = numeric,last_obs_avg = numeric,predict_size = numeric,level = numeric}: Do prediction based on trained AR1-Markov Model.

\item \code{object = ar1_sim_process,last_obs_max = numeric,last_obs_avg = numeric,predict_size = numeric,level = numeric}: Do prediction based on trained AR1 Model.

\item \code{object = markov_sim_process,last_obs_max = numeric,last_obs_avg = numeric,predict_size = numeric,level = numeric}: Do prediction based on trained Markov Model.

\item \code{object = var1_sim_process,last_obs_max = numeric,last_obs_avg = numeric,predict_size = numeric,level = numeric}: Do prediction based on trained VAR Model.
}}

