% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/arima_sim.R, R/auto_arima_sim.R
\name{do_prediction}
\alias{do_prediction}
\alias{do_prediction,arima_sim,list,data.frame-method}
\alias{do_prediction,auto_arima_sim,list,data.frame-method}
\title{Do Prediction}
\usage{
do_prediction(object, trained_result, predict_info)

\S4method{do_prediction}{arima_sim,list,data.frame}(object, trained_result, predict_info)

\S4method{do_prediction}{auto_arima_sim,list,data.frame}(object, trained_result, predict_info)
}
\arguments{
\item{object}{An S4 sim object.}

\item{trained_result}{A list or other class returend by \code{train_model}, containing trained model information.}

\item{predict_info}{A dataframe representing all the past predicted or scheduled information.}
}
\value{
The updated \code{predict_info} on the last row.
}
\description{
This is a generic function that do prediction according to the input object type.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = arima_sim,trained_result = list,predict_info = data.frame}: Do prediction based on trained AR1 Model.

\item \code{object = auto_arima_sim,trained_result = list,predict_info = data.frame}: Do prediction based on trained AR1 Model.
}}

