% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{find_state_num}
\alias{find_state_num}
\title{Find Corresponding State}
\usage{
find_state_num(obs, state_num)
}
\arguments{
\item{obs}{A numeric input of observation.}

\item{state_num}{The total number of states.}
}
\value{
The corresponding state number.
}
\description{
Find the corresponding state for a specific observation with fixed partitioning method.
}
\keyword{internal}
