% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{generate_result}
\alias{generate_result}
\title{Generate The Result Of Simulation}
\usage{
generate_result(object, evaluation, write_result)
}
\arguments{
\item{object}{A subclass of S4 sim object.}

\item{evaluation}{The evaluation dataframe with each row representing each trace, and the columns consists of performance information.}

\item{write_result}{A logical TRUE/FALSE argument to determine whether to store the result of simulation to a file to location stored as an attribute in sim object.}
}
\value{
A list cotaining information needed by S4 sim result object.
}
\description{
Generate The Result Of Simulation
}
\keyword{internal}
