% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ar1_markov_sim.R, R/ar1_sim.R,
%   R/markov_sim.R, R/var1_sim.R
\name{get_character_slots}
\alias{get_character_slots}
\alias{get_character_slots,ar1_markov_sim-method}
\alias{get_character_slots,ar1_sim-method}
\alias{get_character_slots,markov_sim-method}
\alias{get_character_slots,var1_sim-method}
\title{Get The Character Parameters}
\usage{
get_character_slots(object)

\S4method{get_character_slots}{ar1_markov_sim}(object)

\S4method{get_character_slots}{ar1_sim}(object)

\S4method{get_character_slots}{markov_sim}(object)

\S4method{get_character_slots}{var1_sim}(object)
}
\arguments{
\item{object}{An S4 sim object}
}
\value{
A list containing all character parameter informations.

A list containing all character parameter informations.

A list containing all character parameter informations.

A list containing all character parameter informations.
}
\description{
Get The Character Parameters
}
