% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/anovatree_pred.R, R/arima_sim.R,
%   R/auto_arima_sim.R, R/gmm_pred.R, R/markov_sim.R, R/pre_methods.R,
%   R/sim_methods.R, R/surtree_pred.R
\name{get_characteristic_slots}
\alias{get_characteristic_slots}
\alias{get_characteristic_slots,anovatree_pred-method}
\alias{get_characteristic_slots,arima_sim-method}
\alias{get_characteristic_slots,auto_arima_sim-method}
\alias{get_characteristic_slots,gmm_pred-method}
\alias{get_characteristic_slots,markov_sim-method}
\alias{get_characteristic_slots,pred-method}
\alias{get_characteristic_slots,sim-method}
\alias{get_characteristic_slots,surtree_pred-method}
\title{Get The Slots that Are Considered Charactersitics of Simulation}
\usage{
get_characteristic_slots(object)

\S4method{get_characteristic_slots}{anovatree_pred}(object)

\S4method{get_characteristic_slots}{arima_sim}(object)

\S4method{get_characteristic_slots}{auto_arima_sim}(object)

\S4method{get_characteristic_slots}{gmm_pred}(object)

\S4method{get_characteristic_slots}{markov_sim}(object)

\S4method{get_characteristic_slots}{pred}(object)

\S4method{get_characteristic_slots}{sim}(object)

\S4method{get_characteristic_slots}{surtree_pred}(object)
}
\arguments{
\item{object}{An S4 sim or pred object}
}
\value{
A list containing all character parameter informations.

A list containing all character parameter informations.

A list containing all character parameter informations.

A list containing all character parameter informations.

A list containing all character parameter informations.

A list containing all character parameter informations.

A list containing all character parameter informations.

A list containing all character parameter informations.
}
\description{
Get The Slots that Are Considered Charactersitics of Simulation
}
