% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/arima_sim.R, R/sim_methods.R
\name{get_characteristic_slots}
\alias{get_characteristic_slots}
\alias{get_characteristic_slots,arima_sim-method}
\alias{get_characteristic_slots,sim-method}
\title{Get The Slots that Are Considered Charactersitics of Simulation}
\usage{
get_characteristic_slots(object)

\S4method{get_characteristic_slots}{arima_sim}(object)

\S4method{get_characteristic_slots}{sim}(object)
}
\arguments{
\item{object}{An S4 sim object}
}
\value{
A list containing all character parameter informations.

A list containing all character parameter informations.
}
\description{
Get The Slots that Are Considered Charactersitics of Simulation
}
