% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ar1_markov_sim.R, R/ar1_sim.R,
%   R/ari11_sim.R, R/markov_sim.R, R/var1_sim.R, R/vari11_sim.R
\name{get_numeric_slots}
\alias{get_numeric_slots}
\alias{get_numeric_slots,ar1_markov_sim-method}
\alias{get_numeric_slots,ar1_sim-method}
\alias{get_numeric_slots,ari11_sim-method}
\alias{get_numeric_slots,markov_sim-method}
\alias{get_numeric_slots,var1_sim-method}
\alias{get_numeric_slots,vari11_sim-method}
\title{Get The Numeric Parameters}
\usage{
get_numeric_slots(object)

\S4method{get_numeric_slots}{ar1_markov_sim}(object)

\S4method{get_numeric_slots}{ar1_sim}(object)

\S4method{get_numeric_slots}{ari11_sim}(object)

\S4method{get_numeric_slots}{markov_sim}(object)

\S4method{get_numeric_slots}{var1_sim}(object)

\S4method{get_numeric_slots}{vari11_sim}(object)
}
\arguments{
\item{object}{An S4 sim object}
}
\value{
A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.
}
\description{
Get The Numeric Parameters
}
