% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/anovatree_pred.R, R/arima_sim.R,
%   R/auto_arima_sim.R, R/autopilot_sim.R, R/ets_sim.R, R/gmm_pred.R,
%   R/lm_sim.R, R/markov_sim.R, R/multinom_sim.R, R/nn_sim.R, R/pred_methods.R,
%   R/sim_methods.R, R/structal_sim.R, R/surtree_pred.R, R/var_sim.R
\name{get_param_slots}
\alias{get_param_slots}
\alias{get_param_slots,anovatree_pred-method}
\alias{get_param_slots,arima_sim-method}
\alias{get_param_slots,auto_arima_sim-method}
\alias{get_param_slots,autopilot_sim-method}
\alias{get_param_slots,ets_sim-method}
\alias{get_param_slots,gmm_pred-method}
\alias{get_param_slots,lm_sim-method}
\alias{get_param_slots,markov_sim-method}
\alias{get_param_slots,multinom_sim-method}
\alias{get_param_slots,nn_sim-method}
\alias{get_param_slots,pred-method}
\alias{get_param_slots,sim-method}
\alias{get_param_slots,structual_sim-method}
\alias{get_param_slots,surtree_pred-method}
\alias{get_param_slots,var_sim-method}
\title{Get The Slots that Are Considered Hyperparameters of Simulation}
\usage{
get_param_slots(object)

\S4method{get_param_slots}{anovatree_pred}(object)

\S4method{get_param_slots}{arima_sim}(object)

\S4method{get_param_slots}{auto_arima_sim}(object)

\S4method{get_param_slots}{autopilot_sim}(object)

\S4method{get_param_slots}{ets_sim}(object)

\S4method{get_param_slots}{gmm_pred}(object)

\S4method{get_param_slots}{lm_sim}(object)

\S4method{get_param_slots}{markov_sim}(object)

\S4method{get_param_slots}{multinom_sim}(object)

\S4method{get_param_slots}{nn_sim}(object)

\S4method{get_param_slots}{pred}(object)

\S4method{get_param_slots}{sim}(object)

\S4method{get_param_slots}{structual_sim}(object)

\S4method{get_param_slots}{surtree_pred}(object)

\S4method{get_param_slots}{var_sim}(object)
}
\arguments{
\item{object}{An S4 sim or pred object}
}
\value{
A list containing all numeric parameter informations.

A list containing all numeric parameter information.

A list containing all numeric parameter information.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.
}
\description{
Get The Slots that Are Considered Hyperparameters of Simulation
}
