% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/arima_sim.R, R/auto_arima_sim.R,
%   R/markov_sim.R, R/sim_methods.R
\name{get_param_slots}
\alias{get_param_slots}
\alias{get_param_slots,arima_sim-method}
\alias{get_param_slots,auto_arima_sim-method}
\alias{get_param_slots,markov_sim-method}
\alias{get_param_slots,sim-method}
\title{Get The Slots that Are Considered Hyperparameters of Simulation}
\usage{
get_param_slots(object)

\S4method{get_param_slots}{arima_sim}(object)

\S4method{get_param_slots}{auto_arima_sim}(object)

\S4method{get_param_slots}{markov_sim}(object)

\S4method{get_param_slots}{sim}(object)
}
\arguments{
\item{object}{An S4 sim object}
}
\value{
A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.

A list containing all numeric parameter informations.
}
\description{
Get The Slots that Are Considered Hyperparameters of Simulation
}
