% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{machine_survival}
\alias{machine_survival}
\title{Find Outcome of A Scheduled Foreground Job}
\usage{
machine_survival(
  machine_list,
  scheduler_choice,
  actual_cpu,
  actual_time_bin_idx
)
}
\arguments{
\item{machine_list}{A list representing the prediction upper bounds at different bins of machines.}

\item{scheduler_choice}{A numeric number representing the machine id chosen by scheduler.}

\item{actual_cpu}{A numeric number representing requested CPU for background job.}

\item{actual_time_bin_idx}{A numeric number representing which bin the actual time corresponds to.}
}
\value{
A logical value \code{TRUE} if the job survives, \code{FALSE} otherwise.
}
\description{
Find Outcome of A Scheduled Foreground Job
}
\keyword{internal}
