% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{machines_optimal}
\alias{machines_optimal}
\title{Find Optimal Solution For A Foreground Job}
\usage{
machines_optimal(machine_list, actual_time, actual_cpu, scheduler_choice)
}
\arguments{
\item{machine_list}{A list representing the prediction upper bounds at different bins of machines.}

\item{actual_time}{A numeric number representing actual runtime for foreground job.}

\item{actual_cpu}{A numeric number representing requested CPU for ba}

\item{scheduler_choice}{A numeric number representing the machine id chosen by scheduler.}
}
\value{
A list containing optimal background machine and the score corresponding to the choice.
}
\description{
Find Optimal Solution For A Foreground Job
}
\keyword{internal}
