% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{machines_select}
\alias{machines_select}
\title{Select the Best Machine for A Foreground Job}
\usage{
machines_select(machine_list, prob_vec_lst, job_info)
}
\arguments{
\item{machine_list}{A list representing the prediction upper bounds at different bins of machines.}

\item{prob_vec_lst}{A list representing probability vector at different bins.}

\item{job_info}{A list containing requested CPU and clustering info of background job.}
}
\value{
A list containing best background machine and the score corresponding to the choice.
}
\description{
Select the Best Machine for A Foreground Job
}
\keyword{internal}
