% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{machines_select}
\alias{machines_select}
\title{Select the Best Machine for A Foreground Job}
\usage{
machines_select(
  machine_list,
  signal_machines_idx,
  prob_vec_lst,
  job_info,
  cores,
  constraint_prob
)
}
\arguments{
\item{machine_list}{A list representing the prediction upper bounds at different bins of machines.}

\item{signal_machines_idx}{A vector of indices of machines that update their heartbeats.}

\item{prob_vec_lst}{A list representing probability vector at different bins.}

\item{job_info}{A list containing requested CPU and clustering info of background job.}

\item{cores}{A numeric value representing number of cores to compute scores for each machine.}

\item{constraint_prob}{A numeric value representing the cut off for survival probability of a job to be scheduled.}
}
\value{
A list containing best background machine and the score corresponding to the choice.
}
\description{
Select the Best Machine for A Foreground Job
}
\keyword{internal}
