% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot_generated_trace_diagnosis}
\alias{plot_generated_trace_diagnosis}
\title{Plot the Diagnosis of Generated Traces}
\usage{
plot_generated_trace_diagnosis(
  generated_trace,
  max_trace,
  avg_trace,
  orig_rate,
  new_rate,
  trace_name = NULL,
  ...
)
}
\arguments{
\item{generated_trace}{A numeric value representing generated trace.}

\item{max_trace}{A numeric vector that represents the maximum trace with sample length as \code{max_trace} and same sampled rate. Used for fine tuning of the generated trace, or \code{NULL}. Default value is \code{NULL}.}

\item{avg_trace}{A numeric vector that represents the average trace taken in a fixed sampled rate.}

\item{orig_rate}{A numeric positive integer that is typicall smaller that the frequency of \code{avg_trace} and \code{max_trace}.}

\item{new_rate}{A numeric positive integer that is typicall smaller that the frequency of \code{avg_trace} and \code{max_trace}.}

\item{trace_name}{A character or NULL acting as an identifier for the trace.}

\item{...}{Characters that represent the name of parent directories that will be passed to \code{write_location_check}.}
}
\description{
Plot the generated trace and the actual maximum and average to compare.
}
