% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{auto_detect_parameters}
\alias{auto_detect_parameters}
\alias{plot_sim_charwise}
\title{Automatically Detect Parameters in Charwise Summary}
\usage{
auto_detect_parameters(charwise_summ)

plot_sim_charwise(
  charwise_summ,
  mapping = list(shape = "window_size", alpha = "granularity", fill = "react_speed",
    color = "model_num"),
  name,
  ...
)
}
\arguments{
\item{charwise_summ}{A dataframe containing the scores in all parameter settings and their performance.}

\item{mapping}{A named list with keys representing graphical aesthetics parameters and values represetning columnames of \code{charwise_summ} to be mapped to.}

\item{name}{A character that represents the identifier or name of the plot.}

\item{...}{Characters that represent the name of parent directories that will be passed to \code{write_location_check}.}
}
\description{
Only used for charwise summary dataframe, used for creating mapping for aesthetics for platting using \code{plot_sim_charwise}.

Plot charwise result for simulation with each datapoint corresponds to average scores of all traces with one configuration.
}
