% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ar1_markov_sim.R, R/markov_sim.R,
%   R/sim_methods.R
\name{plot_sim_overall}
\alias{plot_sim_overall}
\alias{plot_sim_overall,sim-method}
\title{Plot Simulation Result Type Overall}
\usage{
plot_sim_overall(object)

\S4method{plot_sim_overall}{sim}(object)

\S4method{plot_sim_overall}{sim}(object)

\S4method{plot_sim_overall}{sim}(object)
}
\arguments{
\item{object}{An S4 sim object}
}
\value{
A plot object

A plot object

A plot object
}
\description{
Plot overall result for simulation with each datapoint corresponds to average scores of all traces with one configuration.
}
