% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_sim_overall}
\alias{plot_sim_overall}
\title{Plot Simulation Result Type Overall}
\usage{
plot_sim_overall(object, overall_summ, result_loc)
}
\arguments{
\item{object}{An S4 sim object}

\item{overall_summ}{A dataframe containing the scores in all parameter settings.}

\item{result_loc}{A character that specify the path to which the result of simulations will be saved to.}
}
\description{
Plot overall result for simulation with each datapoint corresponds to average scores of all traces with one configuration.
}
