% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\name{plot_sim_paramwise}
\alias{plot_sim_paramwise}
\alias{plot_sim_paramwise,sim,data.frame,list,character-method}
\title{Plot Simulation Result Type Tracewise}
\usage{
plot_sim_paramwise(object, score, summ, result_loc)

\S4method{plot_sim_paramwise}{sim,data.frame,list,character}(object, score, summ, result_loc)
}
\arguments{
\item{object}{An S4 sim object.}

\item{score}{A dataframe containing the scores for all traces.}

\item{summ}{A vector containing the summary of scores for this parameter setting.}

\item{result_loc}{A character that specify the path to which the result of simulations will be saved to.}
}
\description{
Plot tracewise result for simulation with each plot corresponds to the performance of one single trace.
}
