% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\name{plot_sim_paramwise}
\alias{plot_sim_paramwise}
\alias{plot_sim_paramwise,sim,data.frame,list-method}
\title{Plot Simulation Result Type Tracewise}
\usage{
plot_sim_paramwise(object, score, summ)

\S4method{plot_sim_paramwise}{sim,data.frame,list}(object, score, summ)
}
\arguments{
\item{object}{An S4 sim object.}

\item{score}{A dataframe representing score1 and score2 for each trace.}

\item{summ}{A list containing summary of simulation result.}
}
\description{
Plot tracewise result for simulation with each plot corresponds to the performance of one single trace.
}
