% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_sim_paramwise}
\alias{plot_sim_paramwise}
\title{Plot Simulation Result Type Tracewise}
\usage{
plot_sim_paramwise(object)
}
\arguments{
\item{object}{An S4 sim object.}
}
\description{
Plot tracewise result for simulation with each plot corresponds to the performance of one single trace.
}
