% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot_sim_tracewise}
\alias{plot_sim_tracewise}
\title{Plot Simulation Result Type Tracewise}
\usage{
plot_sim_tracewise(predict_info, name, ...)
}
\arguments{
\item{predict_info}{A dataframe containing all the past predicted information.}

\item{name}{A character that represents the identifier or name of the plot.}

\item{...}{Characters that represent the name of parent directories that will be passed to \code{write_location_check}.}
}
\description{
Plot tracewise result for simulation with each plot corresponds to the performance of one single trace.
}
