% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ar1_sim.R, R/ari11_sim.R,
%   R/sim_methods.R, R/var1_sim.R, R/vari11_sim.R
\name{plot_sim_tracewise}
\alias{plot_sim_tracewise}
\alias{plot_sim_tracewise,ar1_sim,numeric,character,data.frame,data.frame,data.frame,numeric,list-method}
\alias{plot_sim_tracewise,ari11_sim,numeric,character,data.frame,data.frame,data.frame,numeric,list-method}
\alias{plot_sim_tracewise,sim,numeric,character,data.frame,data.frame,data.frame,numeric,list-method}
\alias{plot_sim_tracewise,var1_sim,numeric,character,data.frame,data.frame,data.frame,numeric,list-method}
\alias{plot_sim_tracewise,vari11_sim,numeric,character,data.frame,data.frame,data.frame,numeric,list-method}
\title{Plot Simulation Result Type Tracewise}
\usage{
plot_sim_tracewise(
  object,
  index,
  trace_name,
  trainset,
  testset,
  prev_score,
  last_score,
  decision
)

\S4method{plot_sim_tracewise}{ar1_sim,numeric,character,data.frame,data.frame,data.frame,numeric,list}(
  object,
  index,
  trace_name,
  trainset,
  testset,
  prev_score,
  last_score,
  decision
)

\S4method{plot_sim_tracewise}{ari11_sim,numeric,character,data.frame,data.frame,data.frame,numeric,list}(
  object,
  index,
  trace_name,
  trainset,
  testset,
  prev_score,
  last_score,
  decision
)

\S4method{plot_sim_tracewise}{sim,numeric,character,data.frame,data.frame,data.frame,numeric,list}(
  object,
  index,
  trace_name,
  trainset,
  testset,
  prev_score,
  last_score,
  decision
)

\S4method{plot_sim_tracewise}{var1_sim,numeric,character,data.frame,data.frame,data.frame,numeric,list}(
  object,
  index,
  trace_name,
  trainset,
  testset,
  prev_score,
  last_score,
  decision
)

\S4method{plot_sim_tracewise}{vari11_sim,numeric,character,data.frame,data.frame,data.frame,numeric,list}(
  object,
  index,
  trace_name,
  trainset,
  testset,
  prev_score,
  last_score,
  decision
)
}
\arguments{
\item{object}{An S4 sim object.}

\item{index}{A numeric number representing the index of current parameter setting.}

\item{trace_name}{A character representation of name of current trace.}

\item{trainset}{A dataframe with two columns each representing training set of maximum and training set of average.}

\item{testset}{A dataframe with two columns each representing test set of maximum and test set of average.}

\item{prev_score}{A dataframe with two columns each representing the score 1 and score 2 from past evaluation.s}

\item{last_score}{A numeric vector of length 2 each representing score 1 and score 2 of current evaluation.}

\item{decision}{A list representing the decisions made by current algorithm.}
}
\description{
Plot tracewise result for simulation with each plot corresponds to the performance of one single trace.
}
