% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\name{plot_sim_tracewise}
\alias{plot_sim_tracewise}
\alias{plot_sim_tracewise,sim,data.frame,data.frame,data.frame,numeric,list-method}
\title{Plot Simulation Result Type Tracewise}
\usage{
plot_sim_tracewise(object, trainset, testset, prev_score, last_score, decision)

\S4method{plot_sim_tracewise}{sim,data.frame,data.frame,data.frame,numeric,list}(object, trainset, testset, prev_score, last_score, decision)
}
\arguments{
\item{object}{An S4 sim object.}

\item{trainset}{A dataframe with two columns each representing training set of maximum and training set of average.}

\item{testset}{A dataframe with two columns each representing test set of maximum and test set of average.}

\item{prev_score}{A dataframe with two columns each representing the score 1 and score 2 from past evaluation.s}

\item{last_score}{A numeric vector of length 2 each representing score 1 and score 2 of current evaluation.}

\item{decision}{A list representing the decisions made by current algorithm.}
}
\description{
Plot tracewise result for simulation with each plot corresponds to the performance of one single trace.
}
