% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict_model}
\alias{predict_model}
\title{Schedule Jobs Using Predictions On Given Test Set}
\usage{
predict_model(object_process, testset_max, testset_avg)
}
\arguments{
\item{object_process}{Hidden S4 sim object after training.}

\item{testset_max}{A matrix of size \eqn{n \times m} representing the test set of maximum for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}

\item{testset_avg}{A matrix of size \eqn{n \times m} representing the test set of average for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}
}
\value{
A list containing the resulting prediction informations.
}
\description{
Sequantially schedule jobs using predictions on provided test set.
}
\keyword{internal}
