% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predicting_sim}
\alias{predicting_sim}
\title{Simulation of Scheduling Jobs Based On Predictions.}
\usage{
predicting_sim(
  object,
  x,
  xreg,
  start_point = 1,
  wait_time = 0,
  cores,
  write_type,
  plot_type,
  ...
)
}
\arguments{
\item{object}{A uni-length sim object that represents a specific parameter setting.}

\item{x}{A matrix of size n by m representing the target dataset for scheduling and evaluations.}

\item{xreg}{A matrix of length n by m representing the dataset that target dataset depends on for scheduling and evaluations.}

\item{start_point}{A numeric number that represents the starting point of the simulation. Default value is \code{1}.}

\item{wait_time}{A numeric number that represents the time between training and testing. Default value is \code{0}.}

\item{cores}{The number of threads for parallel programming for multiple traces, not supported for windows users.}

\item{write_type}{A character that represents how to write the result of simulation, can be one of "charwise", "tracewise", "paramwise" or "none".}

\item{plot_type}{A character that represents how to plot the result of simulation can be one of "charwise", "tracewise", "paramwise" or "none".}

\item{...}{Characters that represent the name of parent directories that will be passed to \code{write_location_check}.}
}
\value{
An S4 sim result object.
}
\description{
Sequantially training and testing by scheduling a job.
}
\keyword{internal}
