% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predicting_sim}
\alias{predicting_sim}
\title{Simulation of Scheduling Jobs Based On Predictions With AR1 Model}
\usage{
predicting_sim(object, x, xreg, cores, plot_type, result_loc)
}
\arguments{
\item{object}{A uni-length sim object that represents a specific parameter setting.}

\item{x}{A matrix of size n by m representing the target dataset for scheduling and evaluations.}

\item{xreg}{A matrix of length n by m representing the dataset that target dataset depends on for scheduling and evaluations.}

\item{cores}{The number of threads for parallel programming for multiple traces, not supported for windows users.}

\item{plot_type}{A character that can be one of "overall", "tracewise", "paramwise" or "none".}

\item{result_loc}{A character that specify the path to which the result of simulations will be saved to.}
}
\value{
An S4 sim result object.
}
\description{
Sequantially training and testing by scheduling a job using AR1 Model.
}
\keyword{internal}
