% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_sim.R
\name{predicting_sim_ar1}
\alias{predicting_sim_ar1}
\title{Simulation of Scheduling Jobs Based On Predictions With AR1 Model}
\usage{
predicting_sim_ar1(
  param,
  dataset,
  training_policy,
  schedule_policy,
  adjust_policy,
  mode,
  cores,
  write_result
)
}
\arguments{
\item{param}{A vector containing necessary informations or hyperparameters for AR1 model.}

\item{dataset}{A \eqn{n \times m} matrix, with each column is the time series of maxes and avges of CPU information on a machine.}

\item{training_policy}{\code{"once"} for offline training, \code{"fixed"} for training at fixed time, \code{"dynamic"} for training when previous performance is bad.}

\item{schedule_policy}{\code{"disjoint"} for scheduling at fixed time, \code{"dynamic"} for scheduling again immediately when failed.}

\item{adjust_policy}{\code{TRUE} for "backing off" strategy whenever a mistake is made.}

\item{mode}{\code{"max"} or \code{"avg"} which time series is used as \code{dataset}.}

\item{cores}{The number of threads for parallel programming for multiple traces, not supported for windows users.}

\item{write_result}{TRUE if the result of the experiment is written to a file.}
}
\value{
A dataframe containing the resulting scheduling informations.
}
\description{
Sequantially training and testing by scheduling a job using AR1 Model.
}
