% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\name{result_loc}
\alias{result_loc}
\alias{result_loc<-}
\title{Getter/Setter for Result Location}
\usage{
result_loc(object)

result_loc(object) <- value

\S4method{result_loc}{sim}(object)

\S4method{result_loc}{sim,character}(object) <- value
}
\arguments{
\item{object}{An S4 sim object}

\item{value}{A character value of the path to a directory.}
}
\description{
Result location is a path to a directory you wish to store the result to, the default is the current work directory.
}
