% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_sim}
\alias{run_sim}
\title{Predictions of Background Jobs.}
\usage{
run_sim(
  epoch_setting,
  x,
  xreg,
  start_point = 1,
  wait_time = 0,
  cores = parallel::detectCores(),
  write_type,
  plot_type,
  result_loc = getwd()
)
}
\arguments{
\item{epoch_setting}{A dataframe representing a specific parameter setting.}

\item{x}{A matrix of size n by m representing the target dataset for scheduling and evaluations.}

\item{xreg}{A matrix of length n by m representing the dataset that target dataset depends on for scheduling and evaluations, or \code{NULL}.}

\item{start_point}{A numeric number that represents the starting point of the simulation. Default value is \code{1}.}

\item{wait_time}{A numeric number that represents the time between training and testing. Default value is \code{0}.}

\item{cores}{A numeric numeb representing the number of threads for parallel programming for multiple traces, not supported for windows users.}

\item{write_type}{A character that represents how to write the result of simulation, can be one of "charwise", "tracewise", "paramwise" or "none".}

\item{plot_type}{A character that represents how to plot the result of simulation can be one of "charwise", "tracewise", "paramwise" or "none".}

\item{result_loc}{A character that specify the path to which the result of simulations will be saved to. Default is your work directory.}
}
\value{
A list of S4 sim result object.
}
\description{
Sequantially training and testing by predicting the availability of CPU resource at next windows.
}
