% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_sim}
\alias{run_sim}
\title{Run Simulation}
\usage{
run_sim(
  object,
  dataset_max,
  dataset_avg,
  cpu_required,
  cores = parallel::detectCores(),
  write_result = TRUE,
  plot_type
)
}
\arguments{
\item{object}{An S4 sim object.}

\item{dataset_max}{A matrix of size \eqn{n \times m} representing the dataset of maximum for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}

\item{dataset_avg}{A matrix of size \eqn{n \times m} representing the dataset of average for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}

\item{cpu_required}{A vector of length \eqn{m}, each element is the size of the job trying to be scheduled on corresponding machine, default value is \code{NULL}.}

\item{cores}{The number of threads for parallel programming for multiple traces, not supported for windows users.}

\item{write_result}{TRUE if the result of the experiment is written to a file.}

\item{plot_type}{A character that can be one of "overall", "tracewise", "paramwise" or "none".}
}
\value{
An S4 sim result object.
}
\description{
The simulation dependes on the \code{type} attribute of the sim object, if \code{"scheduling"} is provided, simulation sequantially training and testing by scheduling a job with job size supplied by \code{cpu_required}, if \code{"predicting"} is supplied, sequential training and testing is made by predictions based on previous observations.
}
