% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{run_sim_pred}
\alias{run_sim_pred}
\title{Combinations of Predictions of Background Jobs and Foreground Jobs}
\usage{
run_sim_pred(
  param_setting_sim,
  param_setting_pred,
  background_x,
  background_xreg,
  foreground_x,
  foreground_xreg,
  bins = c(0, 1, 2, 6, 10, 14, 18, 22, 26, 30, 50, 80, 205),
  cores = parallel::detectCores(),
  write_type = "none",
  result_loc = getwd()
)
}
\arguments{
\item{param_setting_sim}{A dataframe representing a specific parameter setting for sim object.}

\item{param_setting_pred}{A dataframe representing a specific parameter setting for pred object.}

\item{background_x}{A matrix of size n by m representing the target dataset for scheduling and evaluations.}

\item{background_xreg}{A matrix of size n by m representing the target dataset for scheduling and evaluations, or \code{NULL}.}

\item{foreground_x}{A matrix of size n by m representing the target dataset for scheduling and evaluations.}

\item{foreground_xreg}{A matrix of size n by m representing the dataset that target dataset depends on for predicting.}

\item{bins}{A numeric vector representing a specific partioning of time}

\item{cores}{A numeric numeb representing the number of threads for parallel programming for multiple traces, not supported for windows users.}
}
\value{
A dataframe containing the decisions made by scheduler and the optimal decision.
}
\description{
Sequantially training and testing by predicting the availability of CPU resource at next windows.
}
