% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sim_pred.R
\name{run_sim_pred}
\alias{run_sim_pred}
\title{Combinations of Predictions of Background Jobs and Foreground Jobs}
\usage{
run_sim_pred(
  param_setting_sim,
  param_setting_pred,
  foreground_x,
  foreground_xreg,
  lag_xreg = TRUE,
  sim_length,
  use_adjustment = FALSE,
  load_foreground_result = NULL,
  background_x,
  background_xreg,
  load_background_result = NULL,
  heartbeats_percent = 1,
  bins = c(0, 1, 2, 6, 10, 14, 18, 22, 26, 30, 50, 80, 205),
  repeats = 10,
  cores = parallel::detectCores(),
  write_type = "none",
  result_loc = getwd()
)
}
\arguments{
\item{param_setting_sim}{A dataframe representing a specific parameter setting for sim object.}

\item{param_setting_pred}{A dataframe representing a specific parameter setting for pred object.}

\item{foreground_x}{A matrix of size n by m representing the target dataset for scheduling and evaluations.}

\item{foreground_xreg}{A matrix of size n by m representing the target dataset for scheduling and evaluations, or \code{NULL}.}

\item{lag_xreg}{A logical value representing whether the reggressor for foreground jobs needs to be lagged before applying the foreground model. Default value is \code{TRUE}.}

\item{sim_length}{A numeric integer representing the length of time for simulation, training size excluded.}

\item{use_adjustment}{A logical value controlling whether adjustment plicy will be used based on the previous foreground predictions, parameter setting \code{react_speed} will be used.}

\item{load_foreground_result}{A character value representing the location of the previously run foregound simulation results. Default value is \code{NULL}.}

\item{background_x}{A matrix of size n by m representing the target dataset for scheduling and evaluations.}

\item{background_xreg}{A matrix of size n by m representing the dataset that target dataset depends on for predicting.}

\item{load_background_result}{A character value representing the location of the previously run foregound simulation results. Default value is \code{NULL}.}

\item{heartbeats_percent}{A numeric value representing the percentage of sampled machines will periodically send "heat beats", which is the availability information to the scheduler. Default value is \code{1}.}

\item{bins}{A numeric vector representing the discretization will be used on background job length, the first value representing the lower bound such as \code{0}, last value representing the upper bound such as \code{205}.}

\item{repeats}{A numeric integer representing the number of repetitions for each setting, the result will be the mean of those repeated results.}

\item{cores}{A numeric value representing the number of threads for parallel programming for multiple traces, not supported for windows users.}

\item{write_type}{A character that represents how to write the result of simulation, can be "detailed", "summary" or "none". Default value is \code{"none"}.}

\item{result_loc}{A character that specify the path to which the result of simulations will be saved to. Default is your work directory.}
}
\value{
A dataframe containing the decisions made by scheduler.
}
\description{
Sequantially training and testing by predicting the availability of CPU resource at next windows.
}
