% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\docType{class}
\name{sim-class}
\alias{sim-class}
\alias{sim}
\alias{ar1_sim-class}
\alias{ar1_sim}
\alias{markov_sim-class}
\alias{markov_sim}
\alias{var1_sim-class}
\alias{var1_sim}
\alias{ar1_markov_sim-class}
\alias{ar1_markov_sim}
\title{An S4 class to represent a simulation object.}
\description{
An S4 class to represent a simulation object.
}
\section{Functions}{
\itemize{
\item \code{ar1_sim-class}: A ar1_sim object that contains S4 sim object, with additional slot \code{mode}
}}

\section{Slots}{

\describe{
\item{\code{name}}{A character that can either be \code{"AR1"}, \code{"Markov"}, \code{"AR1_Markov"}, \code{"VAR1"}, additional slots may need to be supplied for different simulations.}

\item{\code{type}}{A character that specify the type of simulation to run, this can either be \code{"scheduling"} or \code{"predicting"}.}

\item{\code{window_size}}{A numeric vector that can only be integers to specify how many observations to be aggregated as one.}

\item{\code{cut_off_prob}}{A numeric vector that is the the maximum probability allowed to have next scheduling failing if \code{type = "scheduling"}, and the level of the prediction interval if \code{type = "predicting"}. Default values are \code{0.005, 0.01, 0.1}.}

\item{\code{granularity}}{A numeric vector that specify the amount of CPU usage can be scheduled by one core, if \code{0} is provided, then granularity is not considered. Default values are \code{100/32, 100/64, 0}.}

\item{\code{train_size}}{A numeric vector that specify the training size used for simulations. Default values are \code{2000, 3000, 4000}.}

\item{\code{update_freq}}{A numeric vector that specify the length of testing after each training step, also the amount of step to update after testing step is complete. Default values are \code{2000, 3000, 4000}.}

\item{\code{train_policy}}{A character that specify the training policy, this can either be \code{"once"}, \code{"fixed"} and \code{dynamic}. Default value is \code{"fixed"}.}

\item{\code{schedule_policy}}{A character that specify the scheduling policy, this can either be \code{"disjoint"} or \code{"dynamic"}. Default value is \code{"dynamic"}.}

\item{\code{adjust_policy}}{A character that specify the adjustment policy, this can either be \code{"back_off"} or \code{"none"}. Defaut is \code{"none"}.}

\item{\code{result_loc}}{A character that specify the path to which the result of simulations will be saved to. Default is your work directory.}

\item{\code{state_num}}{A numeric vector that specify the number of states for models that require this attribute.}

\item{\code{mode}}{A character that specify the targeting trace to be tested on, this can either be \code{"max"} or \code{"avg"} for max traces and average traces respectively.}
}}

