% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_class.R, R/arima_sim.R, R/auto_arima_sim.R,
%   R/autopilot_sim.R, R/lm_sim.R, R/markov_sim.R, R/nn_sim.R, R/var_sim.R
\docType{class}
\name{sim-class}
\alias{sim-class}
\alias{sim}
\alias{arima_sim-class}
\alias{arima_sim}
\alias{auto_arima_sim-class}
\alias{auto_arima_sim}
\alias{autopilot_sim-class}
\alias{autopilot_sim}
\alias{lm_sim-class}
\alias{lm_sim}
\alias{markov_sim-class}
\alias{markov_sim}
\alias{nn_sim-class}
\alias{nn_sim}
\alias{var_sim-class}
\alias{var_sim}
\title{An S4 Class to Represent A Simulation.}
\arguments{
\item{probability_function}{A function pointer to the function to compute the cdf, used in \code{combin_sim_pred} function. Default value is \code{stats::qnorm}}

\item{probability_function_additional_argument}{A list containing additional arguments supplied to \code{probability_function}. Default value is \code{list()}.}

\item{outlier_type}{A character representing the type of outlier it will be treated, it can be None for not checking outliers, AO as additive outliers, IO as innovative outliers, LS as level shift, or All for taking account of all outlier types. Default value is \code{"None"}.}

\item{outlier_cval}{A numeric value representing the critical value to determine the significance of each type of outlier. If NA_real_ is supplied, then it uses defaults: If n ≤ 50 then cval is set equal to 3.0; If n ≥ 450 then cval is set equal to 4.0; otherwise cval is set equal to 3 + 0.0025 * (n - 50). Default value is \code{NA_real_}.}

\item{outlier_prediction}{A character representing the distribution of occurences of outliers, and the prior distribution for probability of occurences. Current choices are "None", "Categorical" for predictions based on MLE, "Categorical-Dirichlet" for Bayesian prediction and prior distribution.}

\item{outlier_prediction_prior}{A numeric vector representing the starting value of the hyperparameters of prior distirbution.}

\item{outlier_prediction_update_param}{A logical value representing whether to update the value of the hyperparameters of prior distribution or MLE estimations of parameters depending on \code{outlier_prediction}.}

\item{half_life}{A numerc integer representing the number of windows for the weight to drop to half in jth-quantile. Default value is \code{144}.}

\item{breaks}{A numeric integer or vector representing the number of breaks for each histogram in each window or the break points for \code{x}. Passed into \code{hist}. Default value is \code{10}.}

\item{cut_off_weight}{A numeric value that is close to zero, representing the smallest weight possible, lower which the weight will be considered as zero. Default value is \code{0.001}.}

\item{res_dist}{A character representing the distribution of residual, \code{"normal"} for normal distribution, \code{"skew_norm"} for skewed normal distribution, or \code{"empirical"} for empirical distribution. Default value is \code{"normal"}.}

\item{window_size_for_reg}{A numeric value representing the window size to be aggregated.}

\item{window_type_for_reg}{A character value indicating how the windowing operation is executed. The value must be one of \code{c("max", "avg")}.}

\item{state_num}{A numeric number that represents the number of states in Markov chain. Default value is \code{8}.}

\item{cluster_type}{A character that represents how each state is partitioned. It can only be either \code{"fixed"} for fixed partitioning from \code{0} to \code{100}, or \code{"quantile"} for dynamic partitioning from minimum value to maximum value using quantiles. Default value is \code{"fixed"}.}

\item{P}{A numeric integer value representing the number of seasonal lags for the input series of neural network. It will be passed into \code{forecast::nnetar} as \code{P}. Default value is \code{0}.}

\item{size}{A numeric integer value representing the number of parameters in the hidden layer of the neural network. If \code{NA_real_} is supplied, half of the number of input nodes plus 1 will be used. Default value is \code{NA_real_}.}

\item{train_args}{A list representing additional call passed into the training function, \code{forecast::nnetar}. Default value is \code{list("repeats" = 50)}.}

\item{pred_args}{A list representing additional call passed into the prediction function, \code{forecast::forecast.nnetar}. Default value is \code{list("bootstrap" = TRUE, "npaths" = 800)}.}

\item{p}{A numeric value representing the autoregressive order for VAR model. Default value is \code{1}.}
}
\description{
An S4 Class to Represent A Simulation.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character that represents the name of the simulation.}

\item{\code{window_size}}{A numeric number that can only be integers to specify how many observations to be aggregated as one. Default value is \code{12}.}

\item{\code{target}}{A numeric number that is the target score for \code{score1}. Default value is \code{0.01}.}

\item{\code{cut_off_prob}}{A numeric number that is the level of the prediction interval. Default value is \code{0.99}.}

\item{\code{granularity}}{A numeric number that specify the amount of CPU usage can be scheduled by one core, if \code{0} is provided, then granularity is not considered. Default values is \code{0}.}

\item{\code{train_policy}}{A character that represents the type of training policy that can either be \code{"offline"}, \code{"fixed"} or \code{"dynamic"}. Default value is \code{"dynamic"}.}

\item{\code{train_size}}{A numeric number that specify the training size used for simulations. Default values is \code{3000}.}

\item{\code{extrap_step}}{A numeric number that specify the number of steps after aggregated by \code{window_size} to predict into the future. Default value is \code{1}.}

\item{\code{update_freq}}{A numeric number that specify the number of times to predict into the future after each training step after aggregated by \code{window_size}. Default values is \code{3}.}

\item{\code{react_speed}}{A numeric number of length two that specify the number of failed/successfull predictions needed to activate/deactive backing off strategy. Default value is \code{c(1, 1)}.}

\item{\code{response}}{A character that specify the targeting trace to be tested on, this can either be \code{"max"} or \code{"avg"} for max traces and average traces respectively. Default value is \code{"max"}.}
}}

