% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_class.R, R/ar1_sim.R, R/markov_sim.R
\docType{class}
\name{sim_process-class}
\alias{sim_process-class}
\alias{ar1_sim_process-class}
\alias{ar1_sim_process}
\alias{markov_sim_process-class}
\alias{markov_sim_process}
\title{An S4 Class to Represent A Hidden Processing Simulation}
\description{
This Class is not created mannually by user, but produced as the result of generic functions: \code{train_model} and \code{do_prediction}. It should extend the corresponding sim object.
}
\section{Slots}{

\describe{
\item{\code{trained_model}}{An additional slot to \code{sim} object, typically a list to store all the information that could represent a trained model. This information should be used in \code{do_prediction} generic function.}

\item{\code{predict_result}}{An additional slot to \code{sim} object, typically a list to store all the information about predicted information. This information should be used in \code{compute_pi_up} generic function.}
}}

\seealso{
\code{\link{sim-class}} for super class, \code{\link{train_model}} for producing \code{sim_process} object, \code{\link{do_prediction}} and \code{\link{compute_pi_up}} for use of \code{sim_process} object.
}
