% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_class.R
\docType{class}
\name{sim_result-class}
\alias{sim_result-class}
\alias{sim_result}
\title{An S4 Class to Represent A Simulation Result.}
\description{
An S4 Class to Represent A Simulation Result.
}
\section{Slots}{

\describe{
\item{\code{type}}{A character that represents the type of the simulation result, it should be either \code{"test"} for a testing batch, \code{"train"} for training model life time, \code{"trace"} for an entire trace, \code{"param"} for all traces in same parameter setting.}

\item{\code{score1.n}}{A numeric value that represents the performance on score 1 consistent with \code{type}.}

\item{\code{score1.w}}{A numeric value that represents the weight of \code{score1.n}.}

\item{\code{score1_adj.n}}{A numeric value that represents the performance on adjusted score1 consistent with \code{type}.}

\item{\code{score1_adj.w}}{A numeric value that represents the weight of \code{score1_adj.n}.}

\item{\code{score2.n}}{A numeric value that represents the performance on score 2 consistent with \code{type}.}

\item{\code{score2.w}}{A numeric value that represents the weight of \code{score2.w}.}

\item{\code{score2_adj.n}}{A numeric value that represents the performance on adjusted score 2 consistent with \code{type}.}

\item{\code{score2_adj.w}}{A numeric value that represents the weight of \code{score2_adj.n}.}
}}

