% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{svt_predicting_sim}
\alias{svt_predicting_sim}
\title{Simulation of Scheduling Jobs Based On Predictions On A Single Trace With AR1 Model}
\usage{
svt_predicting_sim(ts_num, object, x, xreg = NULL, plot_type, result_loc)
}
\arguments{
\item{ts_num}{The corresponding trace/column in \code{dataset}.}

\item{x}{A matrix of size n by m representing the target dataset for scheduling and evaluations.}

\item{xreg}{A matrix of length n by m representing the dataset that target dataset depends on for scheduling and evaluations.}

\item{plot_type}{A character that can be one of "overall", "tracewise", "paramwise" or "none".}

\item{result_loc}{A character that specify the path to which the result of simulations will be saved to.}
}
\value{
A list containing the resulting prediction informations.
}
\description{
Sequantially training and testing by scheduling jobs based on predictions on a single trace using AR1 Model.
}
\keyword{internal}
