% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_sim.R
\name{svt_predicting_sim_ar1}
\alias{svt_predicting_sim_ar1}
\title{Simulation of Scheduling Jobs Based On Predictions On A Single Trace With AR1 Model}
\usage{
svt_predicting_sim_ar1(
  ts_num,
  dataset,
  train_size,
  window_size,
  update_freq,
  cut_off_prob,
  granularity,
  training_policy,
  tolerance,
  schedule_policy,
  adjust_policy,
  mode
)
}
\arguments{
\item{ts_num}{The corresponding trace/column in \code{dataset}.}

\item{dataset}{A \eqn{n \times m} matrix, with each column is the time series of maxes and avges of CPU information on a machine.}

\item{train_size}{The length of data used for training.}

\item{window_size}{The length of predictions.}

\item{update_freq}{The length of testing on scheduing decision each iteration.}

\item{cut_off_prob}{The level of uncertainty of prediction interval.}

\item{granularity}{The granularity of 100 percent of total cpu.}

\item{training_policy}{\code{"once"} for offline training, \code{"fixed"} for training at fixed time, \code{"dynamic"} for training when previous performance is bad.}

\item{tolerance}{The tolerance level of retrain, the quantile of previous performance.}

\item{schedule_policy}{\code{"disjoint"} for scheduling at fixed time, \code{"dynamic"} for scheduling again immediately when failed.}

\item{adjust_policy}{\code{TRUE} for "backing off" strategy whenever a mistake is made.}

\item{mode}{\code{"max"} or \code{"avg"} which time series is used as \code{dataset}.}
}
\value{
A list containing the resulting scheduling informations.
}
\description{
Sequantially training and testing by scheduling jobs based on predictions on a single trace using AR1 Model.
}
