% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_methods.R
\docType{methods}
\name{tolerance}
\alias{tolerance}
\alias{tolerance<-}
\alias{tolerance,sim-method}
\alias{tolerance<-,sim,numeric-method}
\title{Getter/Setter for Tolerance}
\usage{
tolerance(object)

tolerance(object) <- value

\S4method{tolerance}{sim}(object)

\S4method{tolerance}{sim,numeric}(object) <- value
}
\arguments{
\item{object}{An S4 sim object}

\item{value}{A numeric value that is between 0 and 1, exclusive.}
}
\value{
A numeric vector representing tolerance of simulation.
}
\description{
Tolerance is used for dynamic scheme training policy. It controls the minimum quantile of past evaluation the current evaluation needs to supass, otherwise a retrain signal is sent for step.
}
