% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{train_model}
\alias{train_model}
\title{Train Model}
\usage{
train_model(object, trainset_max, trainset_avg)
}
\arguments{
\item{object}{An S4 sim object.}

\item{trainset_max}{A matrix of size \eqn{n \times m} representing the training set of maximum for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}

\item{trainset_avg}{A matrix of size \eqn{n \times m} representing the training set of average for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}
}
\value{
A hidden S4 sim process object that contains trained result.
}
\description{
This is a generic function that trains model according to the input object type, with additional arguments supplied by attributes of the object.
}
