% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/anovatree_pred.R, R/arima_sim.R,
%   R/auto_arima_sim.R, R/autopilot_sim.R, R/gmm_pred.R, R/lm_sim.R,
%   R/markov_sim.R, R/multinom_sim.R, R/nn_sim.R, R/surtree_pred.R, R/var_sim.R
\name{train_model}
\alias{train_model}
\alias{train_model,anovatree_pred,numeric,data.frame,list-method}
\alias{train_model,arima_sim,matrix,NULL,list-method}
\alias{train_model,arima_sim,matrix,matrix,list-method}
\alias{train_model,arima_sim,matrix,list,list-method}
\alias{train_model,auto_arima_sim,matrix,NULL,list-method}
\alias{train_model,auto_arima_sim,matrix,matrix,list-method}
\alias{train_model,auto_arima_sim,matrix,list,list-method}
\alias{train_model,autopilot_sim,matrix,NULL,list-method}
\alias{train_model,gmm_pred,numeric,data.frame,list-method}
\alias{train_model,lm_sim,matrix,matrix,list-method}
\alias{train_model,lm_sim,matrix,list,list-method}
\alias{train_model,markov_sim,matrix,NULL,list-method}
\alias{train_model,markov_sim,matrix,matrix,list-method}
\alias{train_model,multinom_sim,matrix,list,list-method}
\alias{train_model,nn_sim,matrix,NULL,list-method}
\alias{train_model,nn_sim,matrix,matrix,list-method}
\alias{train_model,nn_sim,matrix,list,list-method}
\alias{train_model,surtree_pred,numeric,data.frame,list-method}
\alias{train_model,var_sim,matrix,matrix,list-method}
\alias{train_model,var_sim,matrix,list,list-method}
\title{Train Model}
\usage{
train_model(object, train_x, train_xreg, trained_model)

\S4method{train_model}{anovatree_pred,numeric,data.frame,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{arima_sim,matrix,`NULL`,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{arima_sim,matrix,matrix,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{arima_sim,matrix,list,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{auto_arima_sim,matrix,`NULL`,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{auto_arima_sim,matrix,matrix,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{auto_arima_sim,matrix,list,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{autopilot_sim,matrix,`NULL`,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{gmm_pred,numeric,data.frame,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{lm_sim,matrix,matrix,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{lm_sim,matrix,list,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{markov_sim,matrix,`NULL`,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{markov_sim,matrix,matrix,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{multinom_sim,matrix,list,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{nn_sim,matrix,`NULL`,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{nn_sim,matrix,matrix,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{nn_sim,matrix,list,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{surtree_pred,numeric,data.frame,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{var_sim,matrix,matrix,list}(object, train_x, train_xreg, trained_model)

\S4method{train_model}{var_sim,matrix,list,list}(object, train_x, train_xreg, trained_model)
}
\arguments{
\item{object}{An S4 sim object.}

\item{train_x}{A numeric of length m representing the training set.}

\item{train_xreg}{A numeric or matrix of length or row number m representing the additional regressors for training.}

\item{trained_model}{A list representing the past trained model to update, can be an empty list.}
}
\value{
A list containing trained result.
}
\description{
This is a generic function that trains model according to the input object type, with additional arguments supplied by attributes of the object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = anovatree_pred,train_x = numeric,train_xreg = data.frame,trained_model = list}: Train ARMA Model specific to anovatree_pred object.

\item \code{object = arima_sim,train_x = matrix,train_xreg = NULL,trained_model = list}: Train ARMA Model specific to arima_sim object.

\item \code{object = arima_sim,train_x = matrix,train_xreg = matrix,trained_model = list}: Train ARMA Model specific to arima_sim object.

\item \code{object = arima_sim,train_x = matrix,train_xreg = list,trained_model = list}: Train ARMA Model specific to arima_sim object.

\item \code{object = auto_arima_sim,train_x = matrix,train_xreg = NULL,trained_model = list}: Train ARMA Model specific to auto_arima_sim object.

\item \code{object = auto_arima_sim,train_x = matrix,train_xreg = matrix,trained_model = list}: Train ARMA Model specific to auto_arima_sim object.

\item \code{object = auto_arima_sim,train_x = matrix,train_xreg = list,trained_model = list}: Train ARMA Model specific to auto_arima_sim object.

\item \code{object = autopilot_sim,train_x = matrix,train_xreg = NULL,trained_model = list}: Train model for autopilot recommender.

\item \code{object = gmm_pred,train_x = numeric,train_xreg = data.frame,trained_model = list}: Train ARMA Model specific to gmm_pred object.

\item \code{object = lm_sim,train_x = matrix,train_xreg = matrix,trained_model = list}: Train LM Model specific to lm_sim object.

\item \code{object = lm_sim,train_x = matrix,train_xreg = list,trained_model = list}: Train LM Model specific to lm_sim object.

\item \code{object = markov_sim,train_x = matrix,train_xreg = NULL,trained_model = list}: Train Markov Model specific to markov_sim object.

\item \code{object = markov_sim,train_x = matrix,train_xreg = matrix,trained_model = list}: Train Markov Model specific to markov_sim object.

\item \code{object = multinom_sim,train_x = matrix,train_xreg = list,trained_model = list}: Train Multinomial Model specific to multinom_sim object.

\item \code{object = nn_sim,train_x = matrix,train_xreg = NULL,trained_model = list}: Train NN Model specific to nn_sim object.

\item \code{object = nn_sim,train_x = matrix,train_xreg = matrix,trained_model = list}: Train NN Model specific to nn_sim object.

\item \code{object = nn_sim,train_x = matrix,train_xreg = list,trained_model = list}: Train NN Model specific to nn_sim object.

\item \code{object = surtree_pred,train_x = numeric,train_xreg = data.frame,trained_model = list}: Train ARMA Model specific to surtree_pred object.

\item \code{object = var_sim,train_x = matrix,train_xreg = matrix,trained_model = list}: Train VAR Model specific to var_sim object.

\item \code{object = var_sim,train_x = matrix,train_xreg = list,trained_model = list}: Train VAR Model specific to var_sim object.
}}

