% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ar1_sim.R, R/markov_sim.R,
%   R/var1_sim.R
\name{train_model}
\alias{train_model}
\alias{train_model,ar1_sim,numeric,numeric-method}
\alias{train_model,markov_sim,numeric,numeric-method}
\alias{train_model,var1_sim,numeric,numeric-method}
\title{Train Model}
\usage{
train_model(object, trainset_max, trainset_avg)

\S4method{train_model}{ar1_sim,numeric,numeric}(object, trainset_max, trainset_avg)

\S4method{train_model}{markov_sim,numeric,numeric}(object, trainset_max, trainset_avg)

\S4method{train_model}{var1_sim,numeric,numeric}(object, trainset_max, trainset_avg)
}
\arguments{
\item{object}{An S4 sim object.}

\item{trainset_max}{A matrix of size \eqn{n \times m} representing the training set of maximum for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}

\item{trainset_avg}{A matrix of size \eqn{n \times m} representing the training set of average for scheduling and evaluations, with the initial amount of test set that equals to window size are from training set.}
}
\value{
A hidden S4 sim process object that contains trained result.
}
\description{
This is a generic function that trains model according to the input object type, with additional arguments supplied by attributes of the object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = ar1_sim,trainset_max = numeric,trainset_avg = numeric}: Train AR1 Model specific to ar1_sim object.

\item \code{object = markov_sim,trainset_max = numeric,trainset_avg = numeric}: Train Markov Model specific to markov_sim object.

\item \code{object = var1_sim,trainset_max = numeric,trainset_avg = numeric}: Train VAR Model specific to var1_sim object.
}}

