% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{write_sim_result}
\alias{write_sim_result}
\title{Write The Result Of Simulation}
\usage{
write_sim_result(summ_df, result_type, name, ...)
}
\arguments{
\item{summ_df}{A dataframe containing the scores in all parameter settings and their performance.}

\item{result_type}{A character that can be one of \code{"charwise"}, \code{"tracewise"}, \code{"paramwise"} or \code{"other"}.}

\item{name}{A character that identifies the name of the result.}

\item{...}{Characters that represent the name of parent directories that will be passed to \code{write_location_check}.}
}
\description{
Write the result of a simulation to csv file.
}
\keyword{internal}
